fix(mcp): await global file servers before first turn
Global file-based MCP servers can finish their startup scan after the first Oz
request starts. The first request then omits tools that appear on later turns.

Track each startup scan in the watcher with one PendingScan mechanism. Emit one
ScanComplete event per scan. Accumulate the auto-started global server set in the
manager so a later AgentDriver does not miss the completion event.

Move MCP startup out of driver.rs into driver/mcp_startup.rs. Route every MCP
wait through one await_model_event helper. Run scan completion and server
readiness against one shared, bounded deadline for the cloud-env path and the
initial-global path. Keep the wait non-fatal. Record outcomes in setup
observability.
