+++ app/src/ai/agent_sdk/driver.rs
@@ -19,7 +19,6 @@ use chrono::{DateTime, Utc};
 use futures::FutureExt as _;
 use futures::channel::oneshot;
 use futures::future::{self, Either, FusedFuture, join_all};
-use handlebars::get_arguments;
 use itertools::Itertools as _;
 use oneshot::{Canceled, Receiver};
 use repo_metadata::local_model::IndexedRepoState;
@@ -31,7 +30,6 @@ use warp_cli::agent::{Harness, OutputFormat, RepositoryHeadOverride};
 use warp_cli::mcp::MCPSpec;
 use warp_cli::share::ShareRequest;
 use warp_cli::skill::SkillSpec;
-use warp_core::execution_mode::AppExecutionMode;
 use warp_core::features::FeatureFlag;
 use warp_core::{safe_debug, safe_error, safe_info};
 use warp_errors::{ErrorExt, register_error, report_error, report_if_error};
@@ -57,7 +55,6 @@ use crate::ai::agent_sdk::driver::harness::{
 use crate::ai::agent_sdk::environment_snapshot::{
     EnvironmentSnapshot, EnvironmentSnapshotReporter,
 };
-use crate::ai::agent_sdk::retry::{is_transient_graphql_or_http_error, with_bounded_retry_using};
 use crate::ai::agent_sdk::setup_observability::{SetupClientEventReporter, SetupStep};
 use crate::ai::ambient_agents::task::HarnessModelConfig;
 use crate::ai::ambient_agents::{
@@ -71,8 +68,8 @@ use crate::ai::blocklist::orchestration_event_streamer::{
 };
 use crate::ai::blocklist::orchestration_events::OrchestrationEventService;
 use crate::ai::blocklist::{
-    BlocklistAIHistoryEvent, BlocklistAIHistoryModel, BlocklistAIPermissions,
-    ConversationStatusUpdate, FinalizeReason, finalize_recording_for_conversation,
+    BlocklistAIHistoryEvent, BlocklistAIHistoryModel, ConversationStatusUpdate, FinalizeReason,
+    finalize_recording_for_conversation,
 };
 use crate::ai::cloud_environments::{
     AmbientAgentEnvironment, CloudAmbientAgentEnvironment, GithubRepo, SourceRepo,
@@ -80,19 +77,11 @@ use crate::ai::cloud_environments::{
 use crate::ai::document::ai_document_model::{AIDocumentModel, AIDocumentModelEvent};
 use crate::ai::execution_profiles::profiles::AIExecutionProfilesModel;
 use crate::ai::llms::{LLMId, LLMPreferences};
-use crate::ai::mcp::file_based_manager::{FileBasedMCPManager, FileBasedMCPManagerEvent};
-use crate::ai::mcp::parsing::{ParsedTemplatableMCPServerResult, normalize_mcp_json, resolve_json};
-use crate::ai::mcp::templatable_manager::TemplatableMCPServerManagerEvent;
-use crate::ai::mcp::{
-    JSONMCPServer, MCPServerState, TemplatableMCPServerInstallation, TemplatableMCPServerManager,
-    VariableType, VariableValue, builtin,
-};
 use crate::ai::skills::{
     SkillManager, SkillWatcher, filter_skills_by_spec, read_skills_from_directories,
     resolve_skill_repos,
 };
 use crate::auth::AuthStateProvider;
-use crate::auth::credentials::Credentials;
 use crate::cloud_object::{CloudObject, CloudObjectLookup as _};
 use crate::send_telemetry_from_app_ctx;
 use crate::server::ids::{ServerId, SyncId};
@@ -101,7 +90,6 @@ use crate::server::server_api::ai::{AIClient, TaskStatusUpdate};
 use crate::server::server_api::harness_support::{
     HarnessSupportClient, ResolvePromptAttachedSkill, ResolvePromptRequest,
 };
-use crate::server::server_api::managed_mcp::ManagedMcpClient;
 use crate::terminal::cli_agent_sessions::plugin_manager::{
     CliAgentPluginManager, plugin_manager_for,
 };
@@ -123,11 +111,13 @@ mod error_classification;
 pub(crate) mod git_credentials;
 pub(crate) mod harness;
 mod harness_output_monitor;
+mod mcp_startup;
 pub(super) mod output;
 mod snapshot;
 pub(crate) mod terminal;
 
 use environment::PrepareEnvironmentError;
+use mcp_startup::MCP_SERVER_STARTUP_TIMEOUT;
 pub(crate) use snapshot::upload_snapshot_for_handoff;
 use terminal::TerminalDriverEvent;
 
@@ -175,24 +165,6 @@ where
     }
 }
 
-/// Warn that an MCP server's `{{secret_name}}` references resolved to nothing.
-/// Logs secret names only; resolved values must never reach a log line.
-fn log_unresolved_secret_refs(
-    installation: &TemplatableMCPServerInstallation,
-    unresolved_secret_names: &[String],
-) {
-    if unresolved_secret_names.is_empty() {
-        return;
-    }
-    log::warn!(
-        "MCP server '{}' references secret(s) that are not available to this run: {}. \
-         Check that each secret exists and is attached to this agent or run.",
-        installation.templatable_mcp_server().name,
-        unresolved_secret_names.join(", ")
-    );
-}
-
-const MCP_SERVER_STARTUP_TIMEOUT: Duration = Duration::from_secs(20);
 const HARNESS_SAVE_INTERVAL: Duration = Duration::from_secs(30);
 /// Delay after the initial exit request before retrying with the harness's
 /// follow-up input (e.g. Claude's confirmation-dialog dismissal). Sent
@@ -207,20 +179,6 @@ const HARNESS_EXIT_FORCE_KILL_DELAY: Duration = Duration::from_secs(14);
 /// Bound on the end-of-run wait for `LocalAgentTaskSyncModel` to finish
 /// delivering queued task status updates before the process may exit.
 const TASK_STATUS_FLUSH_TIMEOUT: Duration = Duration::from_secs(5);
-/// Attempt budget for resolving one managed MCP server's client config
-/// (`ManagedMcpClient::create_managed_mcp_client_config`) in
-/// [`AgentDriver::resolve_mcp_specs_with_local_uuids`].
-///
-/// A transient warp-server 5xx during this call otherwise fails the whole run (see
-/// `AgentDriverError::ManagedMcpResolutionFailed`), so this needs a larger budget than the
-/// shared [`crate::server::retry_strategies::MAX_ATTEMPTS`] default (~2s), which is too short
-/// to ride out even a brief backend blip. Against the shared exponential backoff schedule
-/// (500ms, 1s, 2s, 4s, 8s), 6 attempts sum to ~15.5s nominal (~20s with jitter): more than
-/// double the shortest fully-failed window observed in a real incident (a 503 storm from
-/// dying Cloud Run instances, ~6s with zero successful responses inside a ~30s degraded-
-/// capacity period), while staying well under the ~35s a run had before it was marked
-/// FAILED in that incident.
-const MANAGED_MCP_RESOLVE_MAX_ATTEMPTS: usize = 6;
 /// Timeout for individual harness auth preflight commands.
 const PREFLIGHT_CHECK_TIMEOUT: Duration = Duration::from_secs(30);
 pub(crate) const WARP_DRIVE_SYNC_TIMEOUT: Duration = Duration::from_secs(60);
@@ -256,37 +214,6 @@ pub(crate) const WARP_MESSAGE_LISTENER_STATE_ROOT_ENV: &str = "WARP_MESSAGE_LIST
 const LEGACY_OZ_PARENT_LISTENER_MANAGED_EXTERNALLY_ENV: &str =
     "OZ_PARENT_LISTENER_MANAGED_EXTERNALLY";
 const LEGACY_OZ_PARENT_STATE_ROOT_ENV: &str = "OZ_PARENT_STATE_ROOT";
-
-/// Fixed namespace for [`ephemeral_mcp_installation_id`]. Arbitrary; never reused.
-const EPHEMERAL_MCP_INSTALLATION_NAMESPACE: Uuid = Uuid::from_bytes([
-    0xf9, 0x79, 0x1a, 0x88, 0xff, 0xb7, 0x41, 0x88, 0xa6, 0x39, 0xa7, 0x2d, 0xf2, 0x94, 0x22, 0x3f,
-]);
-
-/// Installation id for an ephemeral MCP server (well-known sentinel or non-local
-/// managed MCP UUID) resolved from a managed MCP client config.
-///
-/// With `task_id` (ambient/cloud runs), the id is deterministic: hashing run id +
-/// spec token + server name means a rebuilt sandbox re-resolves the same server to
-/// the same id. Ids can persist in the model's conversation history across a
-/// rebuild; a random id would go stale and fail as "MCP server not found".
-///
-/// Without `task_id` (local sessions; no rebuilds), ids stay random: hashing only
-/// the spec token would collide across concurrent conversations in the same
-/// process, since `TemplatableMCPServerManager` keys installations by this id
-/// process-wide.
-fn ephemeral_mcp_installation_id(
-    task_id: Option<AmbientAgentTaskId>,
-    spec_token: &str,
-    server_name: &str,
-) -> Uuid {
-    match task_id {
-        Some(task_id) => Uuid::new_v5(
-            &EPHEMERAL_MCP_INSTALLATION_NAMESPACE,
-            format!("{task_id}:{spec_token}:{server_name}").as_bytes(),
-        ),
-        None => Uuid::new_v4(),
-    }
-}
 
 /// Abstraction over how [`IdleTimeoutSender::end_run_after`] waits out its deadline, so tests
 /// can substitute a controllable wait for a real, wall-clock-dependent `thread::sleep` and
@@ -1046,12 +973,6 @@ impl ErrorExt for AgentDriverError {
 }
 register_error!(AgentDriverError);
 
-#[derive(Debug, Default)]
-struct ResolvedMcpSpecs {
-    local_uuids: Vec<Uuid>,
-    ephemeral_installations: Vec<TemplatableMCPServerInstallation>,
-}
-
 impl From<warpui::ModelDropped> for AgentDriverError {
     fn from(_: warpui::ModelDropped) -> Self {
         AgentDriverError::InvalidRuntimeState
@@ -1790,869 +1711,6 @@ impl AgentDriver {
         }
     }
 
-    /// Resolve MCP specs into a map of MCP name to `JSONMCPServer` for use in
-    /// third-party harnesses. Each spec is fully resolved (secrets applied, templates
-    /// rendered) so harnesses can serialize directly into their native config format.
-    async fn resolve_mcp_specs_to_json(
-        specs: &[MCPSpec],
-        secrets: Arc<HashMap<String, ManagedSecretValue>>,
-        managed_mcp_client: Arc<dyn ManagedMcpClient>,
-        foreground: &ModelSpawner<Self>,
-    ) -> Result<HashMap<String, JSONMCPServer>, AgentDriverError> {
-        let resolved_specs = Self::resolve_mcp_specs(specs, managed_mcp_client, foreground).await?;
-
-        let local_uuids = resolved_specs.local_uuids;
-        let mut installations = foreground
-            .spawn(move |_, ctx| -> Result<Vec<_>, AgentDriverError> {
-                let manager = TemplatableMCPServerManager::as_ref(ctx);
-                local_uuids
-                    .iter()
-                    .map(|uuid| {
-                        manager
-                            .get_installed_server(uuid)
-                            .cloned()
-                            .ok_or(AgentDriverError::MCPServerNotFound(*uuid))
-                    })
-                    .collect()
-            })
-            .await??;
-        installations.extend(resolved_specs.ephemeral_installations);
-
-        Self::mcp_installations_to_json(installations, secrets.as_ref())
-    }
-
-    fn mcp_installations_to_json(
-        mut installations: Vec<TemplatableMCPServerInstallation>,
-        secrets: &HashMap<String, ManagedSecretValue>,
-    ) -> Result<HashMap<String, JSONMCPServer>, AgentDriverError> {
-        let mut result = HashMap::new();
-
-        for installation in installations.iter_mut() {
-            let unresolved_secret_names = installation.apply_secrets(secrets);
-            if !unresolved_secret_names.is_empty() {
-                return Err(AgentDriverError::MCPUnresolvedSecrets {
-                    server_name: installation.templatable_mcp_server().name.clone(),
-                    secret_names: unresolved_secret_names,
-                });
-            }
-            let resolved = resolve_json(installation);
-            let servers: HashMap<String, JSONMCPServer> = serde_json::from_str(&resolved)
-                .map_err(|e| AgentDriverError::MCPJsonParseError(e.to_string()))?;
-            result.extend(servers);
-        }
-
-        Ok(result)
-    }
-
-    fn apply_secrets_to_ephemeral_mcp_installations(
-        installations: Vec<TemplatableMCPServerInstallation>,
-        secrets: &HashMap<String, ManagedSecretValue>,
-    ) -> (Vec<TemplatableMCPServerInstallation>, Vec<String>) {
-        let mut ready = Vec::with_capacity(installations.len());
-        let mut failures = Vec::new();
-
-        for mut installation in installations {
-            let unresolved_secret_names = installation.apply_secrets(secrets);
-            if unresolved_secret_names.is_empty() {
-                ready.push(installation);
-                continue;
-            }
-
-            log_unresolved_secret_refs(&installation, &unresolved_secret_names);
-            failures.push(format!(
-                "'{}' was not started: unresolved secret reference(s): {}",
-                installation.templatable_mcp_server().name,
-                unresolved_secret_names.join(", ")
-            ));
-        }
-
-        (ready, failures)
-    }
-
-    /// Resolve MCP specs into local UUIDs and ephemeral installations. UUIDs
-    /// are local-first; only non-local UUIDs call managed MCP GraphQL.
-    async fn resolve_mcp_specs(
-        specs: &[MCPSpec],
-        managed_mcp_client: Arc<dyn ManagedMcpClient>,
-        foreground: &ModelSpawner<Self>,
-    ) -> Result<ResolvedMcpSpecs, AgentDriverError> {
-        let (local_installed_uuids, task_id) = foreground
-            .spawn(|me, ctx| {
-                let local_installed_uuids = TemplatableMCPServerManager::as_ref(ctx)
-                    .get_installed_templatable_servers()
-                    .keys()
-                    .copied()
-                    .collect::<HashSet<_>>();
-                (local_installed_uuids, me.task_id)
-            })
-            .await?;
-
-        Self::resolve_mcp_specs_with_local_uuids(
-            specs,
-            &local_installed_uuids,
-            managed_mcp_client,
-            task_id,
-        )
-        .await
-    }
-
-    async fn resolve_mcp_specs_with_local_uuids(
-        specs: &[MCPSpec],
-        local_installed_uuids: &HashSet<Uuid>,
-        managed_mcp_client: Arc<dyn ManagedMcpClient>,
-        task_id: Option<AmbientAgentTaskId>,
-    ) -> Result<ResolvedMcpSpecs, AgentDriverError> {
-        let mut resolved = ResolvedMcpSpecs::default();
-
-        for spec in specs {
-            match spec {
-                MCPSpec::Uuid(uuid) if local_installed_uuids.contains(uuid) => {
-                    resolved.local_uuids.push(*uuid);
-                }
-                MCPSpec::Uuid(uuid) => {
-                    let client_config = with_bounded_retry_using(
-                        &format!("resolve managed MCP server '{uuid}'"),
-                        MANAGED_MCP_RESOLVE_MAX_ATTEMPTS,
-                        is_transient_graphql_or_http_error,
-                        || managed_mcp_client.create_managed_mcp_client_config(uuid.to_string()),
-                    )
-                    .await
-                    .map_err(|err| {
-                        AgentDriverError::ManagedMcpResolutionFailed {
-                            uid: *uuid,
-                            message: format!("{err:#}"),
-                        }
-                    })?;
-                    let installations = Self::installations_from_managed_client_config_json(
-                        &client_config.mcp_config_json,
-                        task_id,
-                        &uuid.to_string(),
-                    )
-                    .map_err(|err| {
-                        AgentDriverError::ManagedMcpResolutionFailed {
-                            uid: *uuid,
-                            message: err.to_string(),
-                        }
-                    })?;
-                    resolved.ephemeral_installations.extend(installations);
-                }
-                MCPSpec::WellKnown(id) => {
-                    // Backstop for specs created before the flag was disabled
-                    // (e.g. persisted configs): skip rather than resolve.
-                    if !FeatureFlag::WellKnownMcpIds.is_enabled() {
-                        log::warn!(
-                            "Skipping well-known MCP server '{id}': WellKnownMcpIds is disabled"
-                        );
-                        continue;
-                    }
-                    // Well-known MCP ids (e.g. "linear") resolve best-effort:
-                    // the server owns the set of recognized ids, and the
-                    // backing integration may be disconnected or the feature
-                    // disabled between dispatch and run setup — so resolution
-                    // failures skip the server instead of failing the run. A
-                    // transient failure still gets the same retry budget as the
-                    // UUID case first, so a brief backend blip doesn't silently
-                    // drop the server from an otherwise-healthy run.
-                    let client_config = match with_bounded_retry_using(
-                        &format!("resolve well-known MCP server '{id}'"),
-                        MANAGED_MCP_RESOLVE_MAX_ATTEMPTS,
-                        is_transient_graphql_or_http_error,
-                        || managed_mcp_client.create_managed_mcp_client_config(id.clone()),
-                    )
-                    .await
-                    {
-                        Ok(client_config) => client_config,
-                        Err(err) => {
-                            log::warn!("Skipping well-known MCP server '{id}': {err:#}");
-                            continue;
-                        }
-                    };
-                    match Self::installations_from_managed_client_config_json(
-                        &client_config.mcp_config_json,
-                        task_id,
-                        id,
-                    ) {
-                        Ok(installations) => {
-                            resolved.ephemeral_installations.extend(installations);
-                        }
-                        Err(err) => {
-                            log::warn!("Skipping well-known MCP server '{id}': {err}");
-                        }
-                    }
-                }
-                MCPSpec::Json(json_str) => {
-                    resolved
-                        .ephemeral_installations
-                        .extend(Self::installations_from_user_mcp_json(json_str)?);
-                }
-            }
-        }
-
-        Ok(resolved)
-    }
-
-    /// Returns the built-in Factory MCP server installation to attach to this
-    /// run, or `None` when it should not be attached.
-    ///
-    /// Interactive clients (GUI/TUI) attach built-in Warp-hosted servers via
-    /// [`TemplatableMCPServerManager::sync_builtin_servers`], which skips CLI
-    /// agent runs. The driver mirrors the same eligibility rules for its
-    /// run-scoped ephemeral startup path: the `FactoryMcp` feature flag, a
-    /// usable bearer token, and no configured server already named
-    /// `warp-factory` (an explicit configuration wins over the built-in).
-    ///
-    /// The token is pinned into the transport at spawn time and is not
-    /// refreshed mid-run: cloud runs authenticate with API keys, which do not
-    /// rotate, so only Firebase-authenticated local runs that outlive their
-    /// token would see factory tool calls start failing.
-    fn builtin_factory_mcp_for_run(
-        credentials: Option<&Credentials>,
-        taken_server_names: &HashSet<String>,
-    ) -> Option<TemplatableMCPServerInstallation> {
-        if !FeatureFlag::FactoryMcp.is_enabled() {
-            return None;
-        }
-        if taken_server_names.contains(builtin::FACTORY_MCP_SERVER_NAME) {
-            log::info!(
-                "Skipping the built-in Factory MCP server: a server named '{}' is already configured for this run",
-                builtin::FACTORY_MCP_SERVER_NAME
-            );
-            return None;
-        }
-        let token = builtin::builtin_bearer_token(credentials?)?;
-        log::info!("Attaching the built-in Factory MCP server to this agent run");
-        Some(builtin::factory_mcp_installation(&token))
-    }
-
-    fn installations_from_user_mcp_json(
-        json_str: &str,
-    ) -> Result<Vec<TemplatableMCPServerInstallation>, AgentDriverError> {
-        let normalized_json = normalize_mcp_json(json_str)
-            .map_err(|e| AgentDriverError::MCPJsonParseError(e.to_string()))?;
-        let parsed_results = ParsedTemplatableMCPServerResult::from_user_json(&normalized_json)
-            .map_err(|e| AgentDriverError::MCPJsonParseError(e.to_string()))?;
-
-        parsed_results
-            .into_iter()
-            .map(|result| {
-                result
-                    .templatable_mcp_server_installation
-                    .ok_or(AgentDriverError::MCPMissingVariables)
-            })
-            .collect()
-    }
-
-    fn installations_from_managed_client_config_json(
-        json_str: &str,
-        task_id: Option<AmbientAgentTaskId>,
-        spec_token: &str,
-    ) -> Result<Vec<TemplatableMCPServerInstallation>, AgentDriverError> {
-        let normalized_json = normalize_mcp_json(json_str)
-            .map_err(|e| AgentDriverError::MCPJsonParseError(e.to_string()))?;
-        let parsed_results = ParsedTemplatableMCPServerResult::from_user_json(&normalized_json)
-            .map_err(|e| AgentDriverError::MCPJsonParseError(e.to_string()))?;
-
-        parsed_results
-            .into_iter()
-            .map(|result| {
-                let ParsedTemplatableMCPServerResult {
-                    mut templatable_mcp_server,
-                    mut variable_values,
-                    ..
-                } = result;
-
-                // Server-rendered literal values (no `{{...}}` ref) must be preserved verbatim.
-                // Drop them from the template's variable list so `apply_secrets` never sees them —
-                // its implicit key-name matching would otherwise let a colliding local secret
-                // (e.g. one named `Authorization`) overwrite a server-issued proxy header.
-                // They stay in `variable_values`, so `resolve_json` still renders them into the
-                // config.
-                templatable_mcp_server
-                    .template
-                    .variables
-                    .retain(|variable| {
-                        let is_literal = variable_values
-                            .get(&variable.key)
-                            .is_some_and(|v| get_arguments(&v.value).is_empty());
-                        !is_literal
-                    });
-
-                // Remaining variables are explicit `{{...}}` placeholders the client fills from
-                // local secrets via `apply_secrets`. Synthesize a placeholder value for any not
-                // captured from env/headers (e.g. command-arg refs like `--token={{API_TOKEN}}`).
-                for variable in templatable_mcp_server.template.variables.iter() {
-                    variable_values
-                        .entry(variable.key.clone())
-                        .or_insert_with(|| VariableValue {
-                            variable_type: VariableType::Text,
-                            value: format!("{{{{{}}}}}", variable.key),
-                        });
-                }
-
-                let installation_id = ephemeral_mcp_installation_id(
-                    task_id,
-                    spec_token,
-                    &templatable_mcp_server.name,
-                );
-                Ok(TemplatableMCPServerInstallation::new(
-                    installation_id,
-                    templatable_mcp_server,
-                    variable_values,
-                ))
-            })
-            .collect()
-    }
-
-    /// Start MCP servers from profile allowlist for the terminal.
-    fn start_profile_mcp_servers(
-        &self,
-        ctx: &mut ModelContext<Self>,
-    ) -> impl Future<Output = Result<(), AgentDriverError>> + use<> {
-        let terminal_id = self.terminal_driver.as_ref(ctx).terminal_view().id();
-        let permissions = BlocklistAIPermissions::as_ref(ctx);
-        let profile_allowlist = permissions.get_mcp_allowlist(ctx, Some(terminal_id));
-
-        if !profile_allowlist.is_empty() {
-            log::info!(
-                "Starting {} MCP servers allowlisted in profile",
-                profile_allowlist.len()
-            );
-        }
-        self.start_mcp_servers(&profile_allowlist, ctx)
-    }
-
-    fn get_mcp_servers_to_start(
-        &self,
-        uuids: &[uuid::Uuid],
-        ctx: &mut ModelContext<Self>,
-    ) -> Result<HashSet<Uuid>, AgentDriverError> {
-        let templatable_mcp_manager = TemplatableMCPServerManager::handle(ctx);
-
-        let mut servers_to_start: HashSet<Uuid> = HashSet::new();
-
-        for uuid in uuids.iter() {
-            if templatable_mcp_manager
-                .as_ref(ctx)
-                .is_server_active_or_pending(*uuid)
-            {
-                log::debug!("MCP server {uuid} is already active or pending; skipping");
-                continue;
-            } else if templatable_mcp_manager
-                .as_ref(ctx)
-                .get_installed_server(uuid)
-                .is_some()
-            {
-                servers_to_start.insert(*uuid);
-            } else {
-                return Err(AgentDriverError::MCPServerNotFound(*uuid));
-            }
-        }
-
-        Ok(servers_to_start)
-    }
-
-    /// Subscribe to MCP server state changes and wait for every server in
-    /// `servers` (keyed by installation UUID, valued by display name) to reach
-    /// a terminal state (`Running` or `FailedToStart`), up to the configured
-    /// startup timeout.
-    ///
-    /// Returns [`AgentDriverError::MCPStartupFailed`] naming the servers that
-    /// failed to start or were still starting at the deadline. Callers decide
-    /// whether that is fatal (see strict MCP startup handling in
-    /// `run_internal`).
-    ///
-    /// Must be called before the servers are spawned so no state changes are
-    /// missed, and never concurrently with another MCP wait: the driver keeps
-    /// at most one subscription to [`TemplatableMCPServerManager`].
-    fn wait_for_mcp_servers_started(
-        &self,
-        servers: HashMap<Uuid, String>,
-        ctx: &mut ModelContext<Self>,
-    ) -> impl Future<Output = Result<(), AgentDriverError>> + use<> {
-        // If no servers to wait for, complete immediately.
-        if servers.is_empty() {
-            return Either::Right(future::ready(Ok(())));
-        }
-
-        // Stall for user-configured timeout, else 20 seconds (configured in [`AgentDriverOptions`]).
-        let timeout = self.mcp_startup_timeout;
-        let (tx, rx) = oneshot::channel::<()>();
-        let mut tx = Some(tx);
-
-        let pending = Arc::new(Mutex::new(servers));
-        let failed: Arc<Mutex<Vec<String>>> = Arc::new(Mutex::new(Vec::new()));
-        let pending_for_subscription = Arc::clone(&pending);
-        let failed_for_subscription = Arc::clone(&failed);
-
-        let templatable_mcp_manager = TemplatableMCPServerManager::handle(ctx);
-
-        // Clear any stale subscription left behind by a previous wait that
-        // timed out, so it can't tear down this wait's subscription.
-        ctx.unsubscribe_from_model(&templatable_mcp_manager);
-        ctx.subscribe_to_model(&templatable_mcp_manager, move |_me, manager, event, ctx| {
-            let TemplatableMCPServerManagerEvent::StateChanged { uuid, state } = event else {
-                return;
-            };
-            let Ok(mut pending_servers) = pending_for_subscription.lock() else {
-                return;
-            };
-            let Some(name) = pending_servers.get(uuid).cloned() else {
-                // If we receive a state change for a server that we're not waiting for, ignore it.
-                return;
-            };
-            match state {
-                MCPServerState::Running => {
-                    pending_servers.remove(uuid);
-                }
-                MCPServerState::FailedToStart => {
-                    pending_servers.remove(uuid);
-                    let error = TemplatableMCPServerManager::as_ref(ctx)
-                        .get_server_error_message(*uuid)
-                        .map(|message| format!(": {message}"))
-                        .unwrap_or_default();
-                    let detail = format!("'{name}' failed to start{error}");
-                    log::warn!("MCP server {detail}");
-                    if let Ok(mut failed_servers) = failed_for_subscription.lock() {
-                        failed_servers.push(detail);
-                    }
-                }
-                MCPServerState::NotRunning
-                | MCPServerState::Starting
-                | MCPServerState::Authenticating
-                | MCPServerState::ShuttingDown => return,
-            }
-            if pending_servers.is_empty() {
-                log::info!("All requested MCP servers reached a terminal state");
-                if let Some(sender) = tx.take() {
-                    let _ = sender.send(());
-                }
-                ctx.unsubscribe_from_model(&manager);
-            }
-        });
-
-        let spawner = ctx.spawner();
-        Either::Left(async move {
-            let wait_result = rx.with_timeout(timeout).await;
-
-            let mut still_starting: Vec<String> = Vec::new();
-            match wait_result {
-                Ok(Ok(())) => {}
-                Ok(Err(Canceled)) => {
-                    log::error!("Subscription dropped before MCP servers started");
-                    return Err(AgentDriverError::InvalidRuntimeState);
-                }
-                Err(TimeoutError) => {
-                    still_starting = pending
-                        .lock()
-                        .map(|pending_servers| pending_servers.values().cloned().collect())
-                        .unwrap_or_default();
-                    still_starting.sort();
-                    // The subscription is now stale; remove it so it can't
-                    // tear down a later wait's subscription. This completes
-                    // before this future resolves, so it cannot race with a
-                    // subsequent wait.
-                    let _ = spawner
-                        .spawn(|_, ctx| {
-                            let manager = TemplatableMCPServerManager::handle(ctx);
-                            ctx.unsubscribe_from_model(&manager);
-                        })
-                        .await;
-                }
-            }
-
-            let mut details = failed
-                .lock()
-                .map(|failed_servers| failed_servers.clone())
-                .unwrap_or_default();
-            details.sort();
-            details.extend(
-                still_starting
-                    .iter()
-                    .map(|name| format!("'{name}' did not start within {}s", timeout.as_secs())),
-            );
-
-            if details.is_empty() {
-                Ok(())
-            } else {
-                Err(AgentDriverError::MCPStartupFailed { details })
-            }
-        })
-    }
-
-    /// Fold an MCP startup result into `degraded`, propagating any error that
-    /// is fatal regardless of the strict MCP startup setting.
-    fn collect_mcp_degradation(
-        result: Result<(), AgentDriverError>,
-        degraded: &mut Vec<String>,
-    ) -> Result<(), AgentDriverError> {
-        match result {
-            Ok(()) => Ok(()),
-            Err(AgentDriverError::MCPStartupFailed { details }) => {
-                degraded.extend(details);
-                Ok(())
-            }
-            Err(other) => Err(other),
-        }
-    }
-
-    /// Apply strict MCP startup handling to a recorded startup result.
-    ///
-    /// Degraded startup (`MCPStartupFailed`) is fatal only in strict mode.
-    /// Otherwise the run continues without the unavailable servers: the
-    /// degradation is logged and reported as a run status message.
-    async fn handle_mcp_startup_result(
-        result: Result<(), AgentDriverError>,
-        foreground: &ModelSpawner<Self>,
-    ) -> Result<(), AgentDriverError> {
-        let Err(error) = result else {
-            return Ok(());
-        };
-        let AgentDriverError::MCPStartupFailed { details } = &error else {
-            return Err(error);
-        };
-        let details = details.join("; ");
-
-        let strict = foreground.spawn(|me, _| me.strict_mcp_startup).await?;
-        if strict {
-            return Err(error);
-        }
-
-        log::warn!(
-            "MCP startup degraded ({details}); continuing without the unavailable MCP servers"
-        );
-
-        // Surface the degradation on the run itself. The server currently only
-        // persists status messages on terminal state transitions, so this is
-        // best-effort until message-only updates are supported.
-        let (task_id, ai_client) = foreground
-            .spawn(|me, ctx| {
-                (
-                    me.task_id,
-                    ServerApiProvider::as_ref(ctx).get_ai_client().clone(),
-                )
-            })
-            .await?;
-        if let Some(task_id) = task_id {
-            let message = format!(
-                "Warning: some MCP servers were unavailable during startup ({details}); continuing without their tools."
-            );
-            if let Err(err) = ai_client
-                .update_agent_task(
-                    task_id,
-                    None,
-                    None,
-                    None,
-                    Some(TaskStatusUpdate::message(message)),
-                    None,
-                    None,
-                )
-                .await
-            {
-                log::warn!("Failed to report MCP startup warning for task {task_id}: {err:#}");
-            }
-        }
-        Ok(())
-    }
-
-    fn spawn_inactive_servers(
-        &self,
-        servers_to_start: HashSet<Uuid>,
-        ctx: &mut ModelContext<Self>,
-    ) {
-        let templatable_mcp_manager = TemplatableMCPServerManager::handle(ctx);
-        templatable_mcp_manager.update(ctx, |manager, ctx| {
-            for uuid in servers_to_start {
-                manager.spawn_server(uuid, ctx);
-            }
-        });
-    }
-
-    fn start_mcp_servers(
-        &self,
-        uuids: &[uuid::Uuid],
-        ctx: &mut ModelContext<Self>,
-    ) -> impl Future<Output = Result<(), AgentDriverError>> + use<> {
-        let servers_to_start = match self.get_mcp_servers_to_start(uuids, ctx) {
-            Ok(val) => val,
-            Err(e) => {
-                return Either::Right(future::ready(Err(e)));
-            }
-        };
-
-        // If we don't need to start any servers, complete immediately.
-        if servers_to_start.is_empty() {
-            return Either::Right(future::ready(Ok(())));
-        }
-
-        log::info!("Starting {} MCP servers...", servers_to_start.len());
-
-        let named_servers: HashMap<Uuid, String> = {
-            let manager = TemplatableMCPServerManager::as_ref(ctx);
-            servers_to_start
-                .iter()
-                .map(|uuid| {
-                    let name = manager
-                        .get_installed_server(uuid)
-                        .map(|installation| installation.templatable_mcp_server().name.clone())
-                        .unwrap_or_else(|| uuid.to_string());
-                    (*uuid, name)
-                })
-                .collect()
-        };
-        let wait = self.wait_for_mcp_servers_started(named_servers, ctx);
-
-        self.spawn_inactive_servers(servers_to_start, ctx);
-
-        Either::Left(wait)
-    }
-
-    /// Start ephemeral MCP servers from inline JSON specifications.
-    /// These servers are not persisted and exist only for the duration of the agent run.
-    fn start_ephemeral_mcp_servers(
-        &self,
-        installations: Vec<TemplatableMCPServerInstallation>,
-        ctx: &mut ModelContext<Self>,
-    ) -> impl Future<Output = Result<(), AgentDriverError>> + use<> {
-        if installations.is_empty() {
-            return Either::Right(future::ready(Ok(())));
-        }
-        let (installations, mut unresolved_failures) =
-            Self::apply_secrets_to_ephemeral_mcp_installations(installations, &self.secrets);
-
-        if !installations.is_empty() {
-            log::info!("Starting {} ephemeral MCP servers...", installations.len());
-        }
-
-        let named_servers: HashMap<Uuid, String> = installations
-            .iter()
-            .map(|installation| {
-                (
-                    installation.uuid(),
-                    installation.templatable_mcp_server().name.clone(),
-                )
-            })
-            .collect();
-        let wait = self.wait_for_mcp_servers_started(named_servers, ctx);
-
-        // Spawn the ephemeral servers.
-        let templatable_mcp_manager = TemplatableMCPServerManager::handle(ctx);
-        templatable_mcp_manager.update(ctx, move |manager, ctx| {
-            for installation in installations {
-                manager.spawn_cli_ephemeral_server(installation, ctx);
-            }
-        });
-
-        Either::Left(async move {
-            let mut failures = match wait.await {
-                Ok(()) => Vec::new(),
-                Err(AgentDriverError::MCPStartupFailed { details }) => details,
-                Err(other) => return Err(other),
-            };
-            failures.append(&mut unresolved_failures);
-            if failures.is_empty() {
-                Ok(())
-            } else {
-                failures.sort();
-                Err(AgentDriverError::MCPStartupFailed { details: failures })
-            }
-        })
-    }
-
-    /// Subscribe to [`FileBasedMCPManagerEvent::CloudEnvMcpScanComplete`]
-    /// paths and return a receiver that fires with auto-start-requested server UUIDs once every repo
-    /// reports in. Must be called **before** `prepare_environment` so no events are missed.
-    fn setup_file_based_mcp_discovery(
-        &self,
-        expected_repos: Vec<PathBuf>,
-        ctx: &mut ModelContext<Self>,
-    ) -> Receiver<Vec<Uuid>> {
-        let (tx, rx) = oneshot::channel::<Vec<Uuid>>();
-
-        if expected_repos.is_empty() {
-            let _ = tx.send(vec![]);
-            return rx;
-        }
-
-        log::info!(
-            "Waiting for {} cloud environment repo(s) to report back file-based MCP server UUIDs...",
-            expected_repos.len()
-        );
-
-        let mut tx = Some(tx);
-        let mut pending_repos: HashSet<PathBuf> = HashSet::from_iter(expected_repos);
-        let mut collected_wait_uuids = Vec::<Uuid>::new();
-
-        let file_based_mcp_manager = FileBasedMCPManager::handle(ctx);
-
-        ctx.subscribe_to_model(&file_based_mcp_manager, move |_me, manager, event, ctx| {
-            if let FileBasedMCPManagerEvent::CloudEnvMcpScanComplete {
-                repo_path,
-                wait_server_uuids,
-                ..
-            } = event
-                && pending_repos.remove(repo_path) {
-                    collected_wait_uuids.extend(wait_server_uuids.iter().copied());
-                    // If we've received all scan results from all cloud environment repos, send
-                    // back the auto-start-requested UUIDs and begin waiting for initialization.
-                    if pending_repos.is_empty() {
-                        let uuids = collected_wait_uuids.clone();
-                        if let Some(sender) = tx.take() {
-                            log::info!(
-                                "Collected {} auto-started file-based MCP server(s) from cloud environment repos",
-                                uuids.len()
-                            );
-                            let _ = sender.send(uuids);
-                        }
-                        ctx.unsubscribe_from_model(&manager);
-                    }
-                }
-        });
-
-        rx
-    }
-
-    /// Wait for auto-start-requested file-based MCP servers to reach a terminal state
-    /// (`Running` or `FailedToStart`). Non-fatal: always completes without returning an error.
-    ///
-    /// **Sequencing note:** `AgentDriver` supports only one active subscription to
-    /// [`TemplatableMCPServerManager`] at a time. This function, [`Self::start_mcp_servers`],
-    /// and [`Self::start_ephemeral_mcp_servers`] must therefore run sequentially, never
-    /// concurrently.
-    fn wait_for_file_based_mcps_running(
-        &self,
-        uuids: Vec<Uuid>,
-        ctx: &mut ModelContext<Self>,
-    ) -> impl Future<Output = ()> + use<> {
-        // Filter out UUIDs that have already reached a terminal state.
-        let mut pending_uuids: HashSet<Uuid> = {
-            let templatable_manager = TemplatableMCPServerManager::as_ref(ctx);
-            uuids
-                .into_iter()
-                .filter(|uuid| {
-                    !matches!(
-                        templatable_manager.get_server_state(*uuid),
-                        Some(MCPServerState::Running) | Some(MCPServerState::FailedToStart)
-                    )
-                })
-                .collect()
-        };
-
-        if pending_uuids.is_empty() {
-            log::info!("All file-based MCP servers have reached a terminal state; proceeding");
-            return Either::Right(future::ready(()));
-        }
-
-        let pending_state_details = {
-            let templatable_manager = TemplatableMCPServerManager::as_ref(ctx);
-            let file_based_manager = FileBasedMCPManager::as_ref(ctx);
-            Arc::new(Mutex::new(
-                pending_uuids
-                    .iter()
-                    .map(|uuid| {
-                        let server_name = file_based_manager
-                            .get_installation_by_uuid(*uuid)
-                            .map(|installation| installation.templatable_mcp_server().name.clone())
-                            .unwrap_or_else(|| "<unknown>".to_string());
-                        let state = templatable_manager
-                            .get_server_state(*uuid)
-                            .map(|state| format!("{state:?}"))
-                            .unwrap_or_else(|| "no state".to_string());
-                        let error = templatable_manager
-                            .get_server_error_message(*uuid)
-                            .map(|message| format!(", error={message}"))
-                            .unwrap_or_default();
-                        (*uuid, format!("{server_name} ({uuid}): {state}{error}"))
-                    })
-                    .collect::<HashMap<_, _>>(),
-            ))
-        };
-        let file_based_mcp_names = {
-            let file_based_manager = FileBasedMCPManager::as_ref(ctx);
-            pending_uuids
-                .iter()
-                .map(|uuid| {
-                    let server_name = file_based_manager
-                        .get_installation_by_uuid(*uuid)
-                        .map(|installation| installation.templatable_mcp_server().name.clone())
-                        .unwrap_or_else(|| "<unknown>".to_string());
-                    (*uuid, server_name)
-                })
-                .collect::<HashMap<_, _>>()
-        };
-        log::info!(
-            "Waiting for {} file-based MCP server(s) to reach a terminal state",
-            pending_uuids.len()
-        );
-
-        let (tx, rx) = oneshot::channel::<()>();
-        let mut tx = Some(tx);
-
-        let templatable_manager_handle = TemplatableMCPServerManager::handle(ctx);
-        let pending_state_details_for_subscription = Arc::clone(&pending_state_details);
-
-        ctx.subscribe_to_model(
-            &templatable_manager_handle,
-            move |_me, manager, event, ctx| {
-                if let TemplatableMCPServerManagerEvent::StateChanged { uuid, state } = event {
-                    if !pending_uuids.contains(uuid) {
-                        return;
-                    }
-                    let server_name = file_based_mcp_names
-                        .get(uuid)
-                        .map(String::as_str)
-                        .unwrap_or("<unknown>");
-                    let error = TemplatableMCPServerManager::as_ref(ctx)
-                        .get_server_error_message(*uuid)
-                        .map(|message| format!(", error={message}"))
-                        .unwrap_or_default();
-                    if let Ok(mut details) = pending_state_details_for_subscription.lock() {
-                        details.insert(*uuid, format!("{server_name} ({uuid}): {state:?}{error}"));
-                    }
-                    match state {
-                        MCPServerState::Running | MCPServerState::FailedToStart => {
-                            pending_uuids.remove(uuid);
-                            if let Ok(mut details) = pending_state_details_for_subscription.lock() {
-                                details.remove(uuid);
-                            }
-                        }
-                        _ => {
-                            return;
-                        }
-                    }
-                    if pending_uuids.is_empty() {
-                        log::info!(
-                            "All file-based MCP servers reached a terminal state; proceeding"
-                        );
-                        if let Some(sender) = tx.take() {
-                            let _ = sender.send(());
-                        }
-                        ctx.unsubscribe_from_model(&manager);
-                    }
-                }
-            },
-        );
-
-        Either::Left(async move {
-            match rx.with_timeout(MCP_SERVER_STARTUP_TIMEOUT).await {
-                Ok(Ok(())) => {}
-                Ok(Err(Canceled)) => {
-                    log::warn!(
-                        "File-based MCP server readiness subscription dropped early; proceeding"
-                    );
-                }
-                Err(TimeoutError) => {
-                    let pending_details = pending_state_details
-                        .lock()
-                        .map(|details| details.values().cloned().join("; "))
-                        .unwrap_or_else(|_| "<unable to read pending state>".to_string());
-                    log::warn!(
-                        "Timed out waiting for file-based MCP servers to reach a terminal state; proceeding without. Still pending: {pending_details}"
-                    );
-                }
-            }
-        })
-    }
-
     /// Resolve global skill specs and the GitHub repositories that should be cloned for them.
     async fn resolve_global_skills(
         foreground: &ModelSpawner<Self>,
@@ -2963,441 +2021,330 @@ impl AgentDriver {
         );
 
         let setup_span = tracing::info_span!("agent_run_setup", tags.cloud_agent = true);
-        let (
-            setup_events,
-            task_id_for_refresh,
-            ai_client_for_refresh,
-            oidc_strategy_for_refresh,
-        ) = async {
-            let (setup_events, environment_snapshot_reporter) = foreground
-                .spawn(|me, ctx| {
-                    let ai_client = ServerApiProvider::as_ref(ctx).get_ai_client().clone();
-                    let background = ctx.background_executor();
-                    match me.task_id {
-                        Some(task_id) => (
-                            SetupClientEventReporter::new(
-                                task_id,
-                                ai_client.clone(),
-                                background.clone(),
-                            ),
-                            EnvironmentSnapshotReporter::new(task_id, ai_client, background),
-                        ),
-                        None => {
-                            report_error!(
-                                "No task ID found for driver - cannot report client events"
-                            );
-                            (
-                                SetupClientEventReporter::noop(
+        let (setup_events, task_id_for_refresh, ai_client_for_refresh, oidc_strategy_for_refresh) =
+            async {
+                let (setup_events, environment_snapshot_reporter) = foreground
+                    .spawn(|me, ctx| {
+                        let ai_client = ServerApiProvider::as_ref(ctx).get_ai_client().clone();
+                        let background = ctx.background_executor();
+                        match me.task_id {
+                            Some(task_id) => (
+                                SetupClientEventReporter::new(
+                                    task_id,
                                     ai_client.clone(),
                                     background.clone(),
                                 ),
-                                EnvironmentSnapshotReporter::noop(ai_client, background),
-                            )
-                        }
-                    }
-                })
-                .await?;
-
-            foreground
-                .spawn(|me, _| me.check_working_dir())
-                .await?
-                .await?;
-
-        // IMPORTANT: Wait for the terminal session to bootstrap before starting MCP servers.
-        // Some of the initializations are necessary for the MCP servers to start correctly.
-        //
-        // Why: MCP server startup can happen before we actually execute the agent prompt. For
-        // `TransportType::CLIServer` MCPs we currently depend on `AISettings.mcp_execution_path`,
-        // which is populated as part of terminal bootstrap. Waiting for the session bootstrap
-        // here avoids a subtle race where MCP spawn runs with an unset PATH and then the driver
-        // only fails via a timeout.
-        setup_events
-            .record_result(SetupStep::TerminalBootstrap, async {
-                foreground
-                    .spawn(|me, ctx| {
-                        me.terminal_driver
-                            .update(ctx, |driver, _| driver.wait_for_session_bootstrapped())
-                    })
-                    .await?
-                    .await
-                    .map_err(|error| AgentDriverError::BootstrapFailed { error })
-            })
-            .await?;
-
-        // Once the terminal session is bootstrapped, perform cloud provider setup before spawning MCP servers.
-        // MCP servers *may* rely on cloud provider credentials.
-        setup_events
-            .record_result(
-                SetupStep::CloudProviderSetup,
-                Self::setup_cloud_providers(&foreground),
-            )
-            .await?;
-
-        // For the Oz harness only: set up MCP servers, model overrides, and profile information.
-        if matches!(&task.harness, HarnessKind::Oz) {
-            let mcp_specs = task.mcp_specs.clone();
-            let managed_mcp_client = foreground
-                .spawn(|_, ctx| ServerApiProvider::as_ref(ctx).get_managed_mcp_client())
-                .await?;
-
-            let mcp_startup_result = setup_events
-                .record_result(SetupStep::McpServerStartup, async {
-                    let resolved_mcp_specs =
-                        Self::resolve_mcp_specs(&mcp_specs, managed_mcp_client, &foreground)
-                            .await?;
-                    let existing_uuids = resolved_mcp_specs.local_uuids;
-                    let mut ephemeral_installations = resolved_mcp_specs.ephemeral_installations;
-
-                    // Attach the built-in Factory MCP server. Interactive
-                    // clients attach built-ins via
-                    // `TemplatableMCPServerManager::sync_builtin_servers`,
-                    // which skips CLI agent runs, so the driver injects the
-                    // same code-owned installation here, scoped to this run.
-                    let local_uuids = existing_uuids.clone();
-                    let mut taken_server_names: HashSet<String> = ephemeral_installations
-                        .iter()
-                        .map(|installation| installation.templatable_mcp_server().name.clone())
-                        .collect();
-                    let credentials = foreground
-                        .spawn(move |_, ctx| {
-                            let (local_names, builtin_already_active) = {
-                                let manager = TemplatableMCPServerManager::as_ref(ctx);
-                                let local_names = local_uuids
-                                    .iter()
-                                    .filter_map(|uuid| {
-                                        manager.get_installed_server(uuid).map(|installation| {
-                                            installation.templatable_mcp_server().name.clone()
-                                        })
-                                    })
-                                    .collect::<Vec<_>>();
-                                let builtin_already_active = manager.is_server_active_or_pending(
-                                    builtin::FACTORY_MCP_INSTALLATION_UUID,
-                                );
-                                (local_names, builtin_already_active)
-                            };
-                            // Interactive clients (GUI/TUI) attach built-ins
-                            // through `sync_builtin_servers`, under the same
-                            // stable installation UUID. The driver currently
-                            // only runs in SDK mode, where that path never
-                            // spawns, but guard anyway so this injection can
-                            // never double-spawn the built-in if the driver
-                            // is ever hosted in an interactive process.
-                            let builtin_owned_by_manager = builtin_already_active
-                                || AppExecutionMode::as_ref(ctx).can_autostart_mcp_servers();
-                            let auth_state = AuthStateProvider::as_ref(ctx).get().clone();
-                            let credentials = (!builtin_owned_by_manager
-                                && !auth_state.is_anonymous_or_logged_out())
-                            .then(|| auth_state.credentials())
-                            .flatten();
-                            (credentials, local_names)
-                        })
-                        .await
-                        .map(|(credentials, local_names)| {
-                            taken_server_names.extend(local_names);
-                            credentials
-                        })?;
-                    if let Some(installation) =
-                        Self::builtin_factory_mcp_for_run(credentials.as_ref(), &taken_server_names)
-                    {
-                        ephemeral_installations.push(installation);
-                    }
-
-                    log::info!(
-                        "Starting {} existing and {} ephemeral MCP servers",
-                        existing_uuids.len(),
-                        ephemeral_installations.len()
-                    );
-
-                    // Run both startup phases even when one degrades, collecting
-                    // degradation details so non-strict runs can continue with
-                    // whichever servers did start.
-                    let mut degraded = Vec::new();
-                    if !existing_uuids.is_empty() {
-                        let result = foreground
-                            .spawn(move |me, ctx| me.start_mcp_servers(&existing_uuids, ctx))
-                            .await?
-                            .await;
-                        Self::collect_mcp_degradation(result, &mut degraded)?;
-                    }
-                    // Start ephemeral MCP servers from inline JSON specs.
-                    if !ephemeral_installations.is_empty() {
-                        let result = foreground
-                            .spawn(move |me, ctx| {
-                                me.start_ephemeral_mcp_servers(ephemeral_installations, ctx)
-                            })
-                            .await?
-                            .await;
-                        Self::collect_mcp_degradation(result, &mut degraded)?;
-                    }
-                    if degraded.is_empty() {
-                        Ok(())
-                    } else {
-                        Err(AgentDriverError::MCPStartupFailed { details: degraded })
-                    }
-                })
-                .await;
-            Self::handle_mcp_startup_result(mcp_startup_result, &foreground).await?;
-            let profile = task.profile.clone();
-            setup_events
-                .record_result(SetupStep::AgentProfileConfiguration, async {
-                    foreground
-                        .spawn(move |me, ctx| me.configure_terminal(profile, ctx))
-                        .await??;
-                    Ok::<(), AgentDriverError>(())
-                })
-                .await?;
-
-            if let Some(model_id) = task.model.clone() {
-                foreground
-                    .spawn(move |me, ctx| me.set_base_model_override(model_id, ctx))
-                    .await??;
-            }
-
-            let profile_mcp_startup_result = setup_events
-                .record_result(SetupStep::ProfileMcpServerStartup, async {
-                    foreground
-                        .spawn(|me, ctx| me.start_profile_mcp_servers(ctx))
-                        .await?
-                        .await
-                })
-                .await;
-            Self::handle_mcp_startup_result(profile_mcp_startup_result, &foreground).await?;
-        }
-
-        // For all harnesses: wait for the shared session and prepare the environment.
-        setup_events
-            .record_result(SetupStep::SharedSessionEstablishment, async {
-                foreground
-                    .spawn(|me, ctx| {
-                        me.terminal_driver
-                            .update(ctx, |driver, _| driver.wait_for_session_shared())
-                    })
-                    .await?
-                    .await
-            })
-            .await?;
-        let global_skill_resolution = setup_events
-            .record_result(
-                SetupStep::GlobalSkillResolution,
-                Self::resolve_global_skills(&foreground),
-            )
-            .await?;
-        // Clone global skill repos before environment prep can change the
-        // terminal's cwd into a single environment repo.
-        // We do this for all harnesses, so that the skills *may* be discovered by third-party
-        // harnesses if appropriate.
-        setup_events
-            .record_result(
-                SetupStep::GlobalSkillRepoClone,
-                Self::clone_global_skill_repos(&foreground, &global_skill_resolution.repos),
-            )
-            .await?;
-        let mut environment_skill_repos = Vec::new();
-
-        let (
-            environment_opt,
-            additional_source_repos,
-            repository_head_overrides,
-            remove_repository_origins,
-        ) = foreground
-            .spawn(|me, _| {
-                (
-                    me.environment.clone(),
-                    me.additional_source_repos.clone(),
-                    me.repository_head_overrides.clone(),
-                    me.remove_repository_origins,
-                )
-            })
-            .await?;
-        let mut setup_commands = environment_opt
-            .as_ref()
-            .map(|environment| environment.setup_commands.clone())
-            .unwrap_or_default();
-        // The Factory definition checkout is run-scoped: the dispatch decides
-        // whether this run gets one by attaching the clone variables,
-        // independent of which environment the run executes in.
-        environment::prepend_factory_definition_clone(&mut setup_commands);
-        let source_repos = environment::merge_repos_deduped(
-            environment_opt
-                .as_ref()
-                .map(AmbientAgentEnvironment::effective_repos)
-                .unwrap_or_default(),
-            additional_source_repos,
-        )?;
-
-        if environment_opt.is_some() || !source_repos.is_empty() || !setup_commands.is_empty() {
-            log::info!("Loading environment...");
-            environment_skill_repos = source_repos.clone();
-
-            // Subscribe to file-based MCP discovery BEFORE prepare_environment triggers the
-            // pipeline so no CloudEnvMcpScanComplete events are missed.
-            //
-            // File-based MCP discovery is Oz-only.
-            // TODO(REMOTE-1345): handle MCP setup for third-party harnesses.
-            let file_based_discovery_rx = match &task.harness {
-                HarnessKind::Oz => {
-                    let source_repos = source_repos.clone();
-                    Some(
-                        foreground
-                            .spawn(move |me, ctx| {
-                                let expected_repo_paths: Vec<PathBuf> = source_repos
-                                    .iter()
-                                    .map(|repo| me.working_dir.join(&repo.repo))
-                                    .collect();
-                                me.setup_file_based_mcp_discovery(expected_repo_paths, ctx)
-                            })
-                            .await?,
-                    )
-                }
-                HarnessKind::ThirdParty(_) | HarnessKind::Unsupported(_) => None,
-            };
-
-            let harness = task.harness.harness();
-            let setup_events_for_environment = setup_events.clone();
-            let source_repos_for_prepare = source_repos;
-            let prepare_outcome = foreground
-                .spawn(move |me, ctx| {
-                    let working_dir = me.working_dir.clone();
-                    me.terminal_driver.update(ctx, |_, ctx| {
-                        environment::prepare_environment(
-                            working_dir,
-                            false, /* is_sandbox */
-                            harness,
-                            environment::RepositoryPreparationOptions::new(
-                                source_repos_for_prepare,
-                                setup_commands,
-                                repository_head_overrides,
-                                remove_repository_origins,
+                                EnvironmentSnapshotReporter::new(task_id, ai_client, background),
                             ),
-                            setup_events_for_environment,
-                            environment_snapshot_reporter.clone(),
-                            ctx,
-                        )
-                    })
-                })
-                .await?
-                .await
-                .map_err(AgentDriverError::from);
-            if let Err(error) = prepare_outcome {
-                // A broken environment is the case post-failure retention exists for, so this
-                // failure must not take the session down with it on the way out.
-                Self::linger_after_failure(&foreground, "environment_setup", &error).await;
-                return Err(error);
-            }
-
-            if let Some(file_based_discovery_rx) = file_based_discovery_rx {
-                // Await discovery: collect UUIDs of file-based MCP servers that were auto-started
-                // while scanning cloned repos.
-                let wait_uuids = setup_events
-                    .record_value(SetupStep::FileBasedMcpDiscovery, async {
-                        match file_based_discovery_rx
-                            .with_timeout(MCP_SERVER_STARTUP_TIMEOUT)
-                            .await
-                        {
-                            Ok(Ok(uuids)) => uuids,
-                            Ok(Err(Canceled)) => {
-                                log::warn!(
-                                    "File-based MCP discovery subscription dropped early; proceeding without"
+                            None => {
+                                report_error!(
+                                    "No task ID found for driver - cannot report client events"
                                 );
-                                vec![]
-                            }
-                            Err(TimeoutError) => {
-                                log::warn!(
-                                    "Timed out waiting for file-based MCP servers to be parsed; proceeding without"
-                                );
-                                vec![]
+                                (
+                                    SetupClientEventReporter::noop(
+                                        ai_client.clone(),
+                                        background.clone(),
+                                    ),
+                                    EnvironmentSnapshotReporter::noop(ai_client, background),
+                                )
                             }
                         }
                     })
-                    .await;
+                    .await?;
 
-                // Wait for auto-started servers to reach Running (non-fatal: always unblocks).
-                if !wait_uuids.is_empty() {
-                    log::info!(
-                        "Checking readiness for {} auto-started file-based MCP server(s)",
-                        wait_uuids.len()
-                    );
+                foreground
+                    .spawn(|me, _| me.check_working_dir())
+                    .await?
+                    .await?;
+
+                // IMPORTANT: Wait for the terminal session to bootstrap before starting MCP servers.
+                // Some of the initializations are necessary for the MCP servers to start correctly.
+                //
+                // Why: MCP server startup can happen before we actually execute the agent prompt. For
+                // `TransportType::CLIServer` MCPs we currently depend on `AISettings.mcp_execution_path`,
+                // which is populated as part of terminal bootstrap. Waiting for the session bootstrap
+                // here avoids a subtle race where MCP spawn runs with an unset PATH and then the driver
+                // only fails via a timeout.
+                setup_events
+                    .record_result(SetupStep::TerminalBootstrap, async {
+                        foreground
+                            .spawn(|me, ctx| {
+                                me.terminal_driver
+                                    .update(ctx, |driver, _| driver.wait_for_session_bootstrapped())
+                            })
+                            .await?
+                            .await
+                            .map_err(|error| AgentDriverError::BootstrapFailed { error })
+                    })
+                    .await?;
+
+                // Once the terminal session is bootstrapped, perform cloud provider setup before spawning MCP servers.
+                // MCP servers *may* rely on cloud provider credentials.
+                setup_events
+                    .record_result(
+                        SetupStep::CloudProviderSetup,
+                        Self::setup_cloud_providers(&foreground),
+                    )
+                    .await?;
+
+                // For the Oz harness only: set up MCP servers, model overrides, and profile information.
+                if matches!(&task.harness, HarnessKind::Oz) {
+                    let mcp_specs = task.mcp_specs.clone();
+                    let managed_mcp_client = foreground
+                        .spawn(|_, ctx| ServerApiProvider::as_ref(ctx).get_managed_mcp_client())
+                        .await?;
+
+                    let mcp_startup_result = setup_events
+                        .record_result(
+                            SetupStep::McpServerStartup,
+                            Self::start_task_mcp_servers(
+                                &mcp_specs,
+                                managed_mcp_client,
+                                &foreground,
+                            ),
+                        )
+                        .await;
+                    Self::handle_mcp_startup_result(mcp_startup_result, &foreground).await?;
+                    let profile = task.profile.clone();
                     setup_events
-                        .record_result(SetupStep::FileBasedMcpReadiness, async {
+                        .record_result(SetupStep::AgentProfileConfiguration, async {
                             foreground
-                                .spawn(move |me, ctx| {
-                                    me.wait_for_file_based_mcps_running(wait_uuids, ctx)
-                                })
-                                .await?
-                                .await;
+                                .spawn(move |me, ctx| me.configure_terminal(profile, ctx))
+                                .await??;
                             Ok::<(), AgentDriverError>(())
                         })
                         .await?;
+
+                    if let Some(model_id) = task.model.clone() {
+                        foreground
+                            .spawn(move |me, ctx| me.set_base_model_override(model_id, ctx))
+                            .await??;
+                    }
+
+                    let profile_mcp_startup_result = setup_events
+                        .record_result(SetupStep::ProfileMcpServerStartup, async {
+                            foreground
+                                .spawn(|me, ctx| me.start_profile_mcp_servers(ctx))
+                                .await?
+                                .await
+                        })
+                        .await;
+                    Self::handle_mcp_startup_result(profile_mcp_startup_result, &foreground)
+                        .await?;
                 }
-            }
-        } else {
-            environment_snapshot_reporter.report(EnvironmentSnapshot::empty());
-        }
 
-        // Skill loading is Oz-only; third-party harnesses have their own skill systems.
-        if matches!(&task.harness, HarnessKind::Oz) {
-            // Load skills from repos synchronously so the initial message includes them.
-            // File trees are ready after prepare_environment and global skill repo cloning above.
-            let GlobalSkillResolution {
-                specs: global_skill_specs,
-                repos: global_skill_repos,
-            } = global_skill_resolution;
-            setup_events
-                .record_value(
-                    SetupStep::EnvironmentSkillLoading,
-                    Self::load_environment_skills(&foreground, environment_skill_repos),
-                )
-                .await;
-            setup_events
-                .record_value(
-                    SetupStep::GlobalSkillLoading,
-                    Self::load_global_skills(&foreground, global_skill_specs, global_skill_repos),
-                )
-                .await;
-            setup_events
-                .record_value(
-                    SetupStep::SkillsDirsLoading,
-                    Self::load_skills_dirs(&foreground),
-                )
-                .await;
-        }
+                // For all harnesses: wait for the shared session and prepare the environment.
+                setup_events
+                    .record_result(SetupStep::SharedSessionEstablishment, async {
+                        foreground
+                            .spawn(|me, ctx| {
+                                me.terminal_driver
+                                    .update(ctx, |driver, _| driver.wait_for_session_shared())
+                            })
+                            .await?
+                            .await
+                    })
+                    .await?;
+                let global_skill_resolution = setup_events
+                    .record_result(
+                        SetupStep::GlobalSkillResolution,
+                        Self::resolve_global_skills(&foreground),
+                    )
+                    .await?;
+                // Clone global skill repos before environment prep can change the
+                // terminal's cwd into a single environment repo.
+                // We do this for all harnesses, so that the skills *may* be discovered by third-party
+                // harnesses if appropriate.
+                setup_events
+                    .record_result(
+                        SetupStep::GlobalSkillRepoClone,
+                        Self::clone_global_skill_repos(&foreground, &global_skill_resolution.repos),
+                    )
+                    .await?;
+                let mut environment_skill_repos = Vec::new();
 
-        let (task_id_for_refresh, ai_client_for_refresh, oidc_strategy_for_refresh) = foreground
-            .spawn(|me, ctx| {
-                let task_id = if FeatureFlag::GitCredentialRefresh.is_enabled() {
-                    me.task_id.map(|id| id.to_string())
-                } else {
-                    None
-                };
-                let ai_client = ServerApiProvider::as_ref(ctx).get_ai_client().clone();
-                // Capture OidcManaged strategy parameters for the proactive Bedrock credential
-                // refresh loop. Only populated when Bedrock OIDC inference is configured.
-                let oidc_strategy = match ApiKeyManager::handle(ctx)
-                    .as_ref(ctx)
-                    .aws_credentials_refresh_strategy()
-                {
-                    AwsCredentialsRefreshStrategy::OidcManaged {
-                        task_id,
-                        role_arn,
-                        region,
-                    } => task_id
+                let (
+                    environment_opt,
+                    additional_source_repos,
+                    repository_head_overrides,
+                    remove_repository_origins,
+                ) = foreground
+                    .spawn(|me, _| {
+                        (
+                            me.environment.clone(),
+                            me.additional_source_repos.clone(),
+                            me.repository_head_overrides.clone(),
+                            me.remove_repository_origins,
+                        )
+                    })
+                    .await?;
+                let mut setup_commands = environment_opt
+                    .as_ref()
+                    .map(|environment| environment.setup_commands.clone())
+                    .unwrap_or_default();
+                // The Factory definition checkout is run-scoped: the dispatch decides
+                // whether this run gets one by attaching the clone variables,
+                // independent of which environment the run executes in.
+                environment::prepend_factory_definition_clone(&mut setup_commands);
+                let source_repos = environment::merge_repos_deduped(
+                    environment_opt
                         .as_ref()
-                        .map(|tid| (tid.clone(), role_arn.clone(), region.clone())),
-                    AwsCredentialsRefreshStrategy::LocalChain => None,
-                };
-                (task_id, ai_client, oidc_strategy)
-            })
-            .await?;
+                        .map(AmbientAgentEnvironment::effective_repos)
+                        .unwrap_or_default(),
+                    additional_source_repos,
+                )?;
 
-            Ok::<_, AgentDriverError>((
-                setup_events,
-                task_id_for_refresh,
-                ai_client_for_refresh,
-                oidc_strategy_for_refresh,
-            ))
-        }
-        .instrument(setup_span)
-        .await?;
+                if environment_opt.is_some()
+                    || !source_repos.is_empty()
+                    || !setup_commands.is_empty()
+                {
+                    log::info!("Loading environment...");
+                    environment_skill_repos = source_repos.clone();
+
+                    // Subscribe to file-based MCP discovery BEFORE prepare_environment triggers the
+                    // pipeline so no CloudEnvMcpScanComplete events are missed.
+                    //
+                    // File-based MCP discovery is Oz-only.
+                    // TODO(REMOTE-1345): handle MCP setup for third-party harnesses.
+                    let cloud_env_mcp_scan = match &task.harness {
+                        HarnessKind::Oz => {
+                            let source_repos = source_repos.clone();
+                            Some(
+                                foreground
+                                    .spawn(move |me, ctx| {
+                                        let expected_repo_paths: Vec<PathBuf> = source_repos
+                                            .iter()
+                                            .map(|repo| me.working_dir.join(&repo.repo))
+                                            .collect();
+                                        me.wait_for_cloud_env_file_based_mcp_scan(
+                                            expected_repo_paths,
+                                            MCP_SERVER_STARTUP_TIMEOUT,
+                                            ctx,
+                                        )
+                                    })
+                                    .await?,
+                            )
+                        }
+                        HarnessKind::ThirdParty(_) | HarnessKind::Unsupported(_) => None,
+                    };
+
+                    let harness = task.harness.harness();
+                    let setup_events_for_environment = setup_events.clone();
+                    let source_repos_for_prepare = source_repos;
+                    let prepare_outcome = foreground
+                        .spawn(move |me, ctx| {
+                            let working_dir = me.working_dir.clone();
+                            me.terminal_driver.update(ctx, |_, ctx| {
+                                environment::prepare_environment(
+                                    working_dir,
+                                    false, /* is_sandbox */
+                                    harness,
+                                    environment::RepositoryPreparationOptions::new(
+                                        source_repos_for_prepare,
+                                        setup_commands,
+                                        repository_head_overrides,
+                                        remove_repository_origins,
+                                    ),
+                                    setup_events_for_environment,
+                                    environment_snapshot_reporter.clone(),
+                                    ctx,
+                                )
+                            })
+                        })
+                        .await?
+                        .await
+                        .map_err(AgentDriverError::from);
+                    if let Err(error) = prepare_outcome {
+                        // A broken environment is the case post-failure retention exists for, so this
+                        // failure must not take the session down with it on the way out.
+                        Self::linger_after_failure(&foreground, "environment_setup", &error).await;
+                        return Err(error);
+                    }
+
+                    if let Some(scan) = cloud_env_mcp_scan {
+                        Self::await_file_based_mcp_startup(
+                            SetupStep::FileBasedMcpDiscovery,
+                            SetupStep::FileBasedMcpReadiness,
+                            scan,
+                            MCP_SERVER_STARTUP_TIMEOUT,
+                            &setup_events,
+                            &foreground,
+                        )
+                        .await?;
+                    }
+                } else {
+                    environment_snapshot_reporter.report(EnvironmentSnapshot::empty());
+                }
+
+                // Skill loading is Oz-only; third-party harnesses have their own skill systems.
+                if matches!(&task.harness, HarnessKind::Oz) {
+                    // Load skills from repos synchronously so the initial message includes them.
+                    // File trees are ready after prepare_environment and global skill repo cloning above.
+                    let GlobalSkillResolution {
+                        specs: global_skill_specs,
+                        repos: global_skill_repos,
+                    } = global_skill_resolution;
+                    setup_events
+                        .record_value(
+                            SetupStep::EnvironmentSkillLoading,
+                            Self::load_environment_skills(&foreground, environment_skill_repos),
+                        )
+                        .await;
+                    setup_events
+                        .record_value(
+                            SetupStep::GlobalSkillLoading,
+                            Self::load_global_skills(
+                                &foreground,
+                                global_skill_specs,
+                                global_skill_repos,
+                            ),
+                        )
+                        .await;
+                    setup_events
+                        .record_value(
+                            SetupStep::SkillsDirsLoading,
+                            Self::load_skills_dirs(&foreground),
+                        )
+                        .await;
+                }
+
+                let (task_id_for_refresh, ai_client_for_refresh, oidc_strategy_for_refresh) =
+                    foreground
+                        .spawn(|me, ctx| {
+                            let task_id = if FeatureFlag::GitCredentialRefresh.is_enabled() {
+                                me.task_id.map(|id| id.to_string())
+                            } else {
+                                None
+                            };
+                            let ai_client = ServerApiProvider::as_ref(ctx).get_ai_client().clone();
+                            // Capture OidcManaged strategy parameters for the proactive Bedrock credential
+                            // refresh loop. Only populated when Bedrock OIDC inference is configured.
+                            let oidc_strategy = match ApiKeyManager::handle(ctx)
+                                .as_ref(ctx)
+                                .aws_credentials_refresh_strategy()
+                            {
+                                AwsCredentialsRefreshStrategy::OidcManaged {
+                                    task_id,
+                                    role_arn,
+                                    region,
+                                } => task_id
+                                    .as_ref()
+                                    .map(|tid| (tid.clone(), role_arn.clone(), region.clone())),
+                                AwsCredentialsRefreshStrategy::LocalChain => None,
+                            };
+                            (task_id, ai_client, oidc_strategy)
+                        })
+                        .await?;
+
+                Ok::<_, AgentDriverError>((
+                    setup_events,
+                    task_id_for_refresh,
+                    ai_client_for_refresh,
+                    oidc_strategy_for_refresh,
+                ))
+            }
+            .instrument(setup_span)
+            .await?;
 
         // Run the harness with a prompt, racing it against optional background refresh
         // loops for git credentials and Bedrock OIDC credentials via
@@ -3405,6 +2352,26 @@ impl AgentDriver {
         // they are dropped automatically when the harness result resolves.
         match task.harness {
             HarnessKind::Oz => {
+                // Join global file-based MCP startup immediately before the first Oz turn so it
+                // can overlap preceding setup without omitting tools from the first request.
+                let initial_global_scan = foreground
+                    .spawn(|me, ctx| {
+                        me.wait_for_initial_global_file_based_mcp_scan(
+                            MCP_SERVER_STARTUP_TIMEOUT,
+                            ctx,
+                        )
+                    })
+                    .await?;
+                Self::await_file_based_mcp_startup(
+                    SetupStep::InitialGlobalMcpScan,
+                    SetupStep::InitialGlobalMcpReadiness,
+                    initial_global_scan,
+                    MCP_SERVER_STARTUP_TIMEOUT,
+                    &setup_events,
+                    &foreground,
+                )
+                .await?;
+
                 let status_rx = foreground
                     .spawn(move |me, ctx| me.execute_run(task.prompt, ctx))
                     .await?;
