+++ app/src/ai/agent_sdk/driver_tests.rs
@@ -1,32 +1,25 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::ffi::OsString;
 use std::fs;
 use std::path::Path;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::Ordering;
 use std::time::Duration;
 
 use chrono::Local;
 use cloud_object_models::CodeForge;
 use futures::channel::oneshot;
 use futures::executor::block_on;
-use http::StatusCode;
 use repo_metadata::{DirectoryWatcher, RepoMetadataEvent, RepoMetadataModel, RepositoryIdentifier};
 use tempfile::TempDir;
 use warp_cli::agent::Harness;
-use warp_cli::mcp::MCPSpec;
 use warp_cli::skill::SkillSpec;
 use warp_cli::{
     OZ_CLI_ENV, OZ_HARNESS_ENV, OZ_PARENT_RUN_ID_ENV, OZ_RUN_ID_ENV, SERVER_ROOT_URL_OVERRIDE_ENV,
     SESSION_SHARING_SERVER_URL_OVERRIDE_ENV, WS_SERVER_URL_OVERRIDE_ENV,
 };
 use warp_core::channel::ChannelState;
-use warp_core::features::FeatureFlag;
 use warp_graphql::ai::AgentTaskState;
-use warp_graphql::mutations::create_managed_mcp_client_config::{
-    CreateManagedMcpClientConfigOutput, ManagedMcpTransportKind,
-};
-use warp_graphql::response_context::ResponseContext;
 use warp_managed_secrets::ManagedSecretValue;
 use warp_multi_agent_api::response_event;
 use warp_util::standardized_path::StandardizedPath;
@@ -34,9 +27,8 @@ use warpui::r#async::Timer;
 use warpui::{App, SingletonEntity as _};
 
 use super::{
-    AgentDriver, AgentDriverError, AgentRunPrompt, CLIAgentSessionStatus, DebugWindowController,
-    IdleTimeoutSender, LEGACY_OZ_PARENT_LISTENER_MANAGED_EXTERNALLY_ENV,
-    LEGACY_OZ_PARENT_STATE_ROOT_ENV, MANAGED_MCP_RESOLVE_MAX_ATTEMPTS,
+    AgentDriver, AgentRunPrompt, CLIAgentSessionStatus, DebugWindowController, IdleTimeoutSender,
+    LEGACY_OZ_PARENT_LISTENER_MANAGED_EXTERNALLY_ENV, LEGACY_OZ_PARENT_STATE_ROOT_ENV,
     OZ_MESSAGE_LISTENER_MANAGED_EXTERNALLY_ENV, OZ_MESSAGE_LISTENER_STATE_ROOT_ENV,
     PlatformErrorCode, SDKConversationOutputStatus, WARP_MESSAGE_LISTENER_STATE_ROOT_ENV,
     build_secret_env_vars, debug_turn_task_state, idle_window_for_cli_session_status,
@@ -59,766 +51,9 @@ use crate::ai::blocklist::{
 };
 use crate::ai::cloud_environments::{GithubRepo, SourceRepo};
 use crate::ai::llms::LLMId;
-use crate::ai::mcp::JSONTransportType;
-use crate::ai::mcp::builtin::{FACTORY_MCP_INSTALLATION_UUID, FACTORY_MCP_SERVER_NAME};
-use crate::ai::mcp::parsing::normalize_mcp_json;
 use crate::ai::skills::SkillManager;
-use crate::auth::credentials::Credentials;
-use crate::server::graphql::GraphQLError;
-use crate::server::server_api::managed_mcp::MockManagedMcpClient;
 use crate::test_util::assert_eventually;
 use crate::test_util::terminal::{add_window_with_terminal, initialize_app_for_terminal_view};
-
-#[test]
-fn test_normalize_single_cli_server() {
-    let input = r#"{"command": "npx", "args": ["-y", "mcp-server"]}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    // Should wrap with a generated name
-    let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-    let parsed = parsed.as_object().unwrap();
-    assert_eq!(parsed.len(), 1);
-    let (_name, server) = parsed.iter().next().unwrap();
-    assert_eq!(server["command"].as_str().unwrap(), "npx");
-}
-
-#[test]
-fn test_normalize_single_sse_server() {
-    let input = r#"{"url": "http://localhost:3000/mcp", "headers": {"API_KEY": "value"}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    // Should wrap with a generated name
-    let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-    let parsed = parsed.as_object().unwrap();
-    assert_eq!(parsed.len(), 1);
-    let (_name, server) = parsed.iter().next().unwrap();
-    assert_eq!(server["url"].as_str().unwrap(), "http://localhost:3000/mcp");
-}
-
-#[test]
-fn test_normalize_already_wrapped_server() {
-    let input = r#"{"my-server": {"command": "npx", "args": []}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    // Should return as-is (no command/url at top level)
-    assert_eq!(result, input);
-}
-
-#[test]
-fn test_normalize_mcp_servers_wrapper() {
-    let input = r#"{"mcpServers": {"server-name": {"command": "npx", "args": []}}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    // Should return as-is (no command/url at top level)
-    assert_eq!(result, input);
-}
-
-#[test]
-fn test_normalize_servers_wrapper() {
-    let input = r#"{"servers": {"server-name": {"url": "http://example.com"}}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    // Should return as-is (no command/url at top level)
-    assert_eq!(result, input);
-}
-
-#[test]
-fn test_normalize_invalid_json() {
-    let input = "not valid json";
-    let result = normalize_mcp_json(input);
-
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_normalize_cli_server_with_env() {
-    let input = r#"{"command": "npx", "args": ["-y", "mcp-server"], "env": {"API_KEY": "secret"}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-    let parsed = parsed.as_object().unwrap();
-    assert_eq!(parsed.len(), 1);
-    let (_name, server) = parsed.iter().next().unwrap();
-    assert_eq!(server["env"]["API_KEY"].as_str().unwrap(), "secret");
-}
-
-#[test]
-fn test_normalize_sse_server_with_headers() {
-    let input =
-        r#"{"url": "http://localhost:5000/mcp", "headers": {"Authorization": "Bearer token"}}"#;
-    let result = normalize_mcp_json(input).unwrap();
-
-    let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-    let parsed = parsed.as_object().unwrap();
-    assert_eq!(parsed.len(), 1);
-    let (_name, server) = parsed.iter().next().unwrap();
-    assert_eq!(
-        server["headers"]["Authorization"].as_str().unwrap(),
-        "Bearer token"
-    );
-}
-
-fn managed_client_config_output(mcp_config_json: &str) -> CreateManagedMcpClientConfigOutput {
-    CreateManagedMcpClientConfigOutput {
-        transport_kind: ManagedMcpTransportKind::Command,
-        mcp_config_json: mcp_config_json.to_string(),
-        proxy_url: None,
-        proxy_token: None,
-        authorization_header_name: None,
-        authorization_header_value: None,
-        expires_at: None,
-        response_context: ResponseContext {
-            server_version: None,
-        },
-    }
-}
-
-fn raw_secret(value: &str) -> ManagedSecretValue {
-    ManagedSecretValue::RawValue {
-        value: value.to_string(),
-    }
-}
-
-fn render_installations(
-    installations: Vec<crate::ai::mcp::TemplatableMCPServerInstallation>,
-    secrets: HashMap<String, ManagedSecretValue>,
-) -> HashMap<String, crate::ai::mcp::JSONMCPServer> {
-    AgentDriver::mcp_installations_to_json(installations, &secrets).unwrap()
-}
-
-#[test]
-fn managed_resolver_local_uuid_does_not_call_managed_client() {
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let mock = MockManagedMcpClient::new();
-    let local_installed_uuids = HashSet::from([uuid]);
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &local_installed_uuids,
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert_eq!(resolved.local_uuids, vec![uuid]);
-    assert!(resolved.ephemeral_installations.is_empty());
-}
-
-#[test]
-fn managed_resolver_non_local_uuid_calls_managed_client() {
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let config_json =
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","env":{"API_TOKEN":"{{API_TOKEN}}"}}}}"#;
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(1)
-        .returning(move |requested_uid| {
-            assert_eq!(requested_uid, uuid.to_string());
-            Ok(managed_client_config_output(config_json))
-        });
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert!(resolved.local_uuids.is_empty());
-    assert_eq!(resolved.ephemeral_installations.len(), 1);
-}
-
-#[test]
-fn well_known_spec_resolves_via_managed_client() {
-    let _flag = FeatureFlag::WellKnownMcpIds.override_enabled(true);
-    let config_json = r#"{"mcpServers":{"linear":{"url":"https://app.warp.dev/mcp/integration-proxy/linear","headers":{"Authorization":"Bearer tok"}}}}"#;
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(1)
-        .returning(move |requested_uid| {
-            assert_eq!(requested_uid, "linear");
-            Ok(managed_client_config_output(config_json))
-        });
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::WellKnown("linear".to_string())],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert!(resolved.local_uuids.is_empty());
-    assert_eq!(resolved.ephemeral_installations.len(), 1);
-}
-
-#[test]
-fn well_known_resolution_failure_skips_server() {
-    let _flag = FeatureFlag::WellKnownMcpIds.override_enabled(true);
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(1)
-        .returning(|_| Err(anyhow::anyhow!("Linear is not connected for this team")));
-
-    // Well-known references are server-injected and best-effort: a failure must
-    // skip the server, not fail run setup.
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::WellKnown("linear".to_string())],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert!(resolved.local_uuids.is_empty());
-    assert!(resolved.ephemeral_installations.is_empty());
-}
-
-#[test]
-fn well_known_resolution_failure_does_not_drop_other_specs() {
-    let _flag = FeatureFlag::WellKnownMcpIds.override_enabled(true);
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let config_json =
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","env":{"API_TOKEN":"{{API_TOKEN}}"}}}}"#;
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(2)
-        .returning(move |requested_uid| {
-            if requested_uid == "linear" {
-                Err(anyhow::anyhow!("Linear is not connected for this team"))
-            } else {
-                assert_eq!(requested_uid, uuid.to_string());
-                Ok(managed_client_config_output(config_json))
-            }
-        });
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[
-            MCPSpec::WellKnown("linear".to_string()),
-            MCPSpec::Uuid(uuid),
-        ],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert_eq!(resolved.ephemeral_installations.len(), 1);
-}
-
-#[test]
-fn well_known_spec_is_skipped_when_flag_disabled() {
-    let _flag = FeatureFlag::WellKnownMcpIds.override_enabled(false);
-    // The managed client must not be called for well-known specs when the
-    // feature is disabled (e.g. a persisted config from a dogfood build).
-    let mock = MockManagedMcpClient::new();
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::WellKnown("linear".to_string())],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert!(resolved.local_uuids.is_empty());
-    assert!(resolved.ephemeral_installations.is_empty());
-}
-
-#[test]
-fn managed_command_config_env_placeholder_uses_local_secret() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","env":{"API_TOKEN":"{{API_TOKEN}}"}}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(
-        installations,
-        HashMap::from([("API_TOKEN".to_string(), raw_secret("real"))]),
-    );
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::CLIServer { env, .. } => {
-            assert_eq!(env.get("API_TOKEN").map(String::as_str), Some("real"));
-        }
-        other => panic!("expected CLI server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_command_config_arg_placeholder_uses_local_secret() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","args":["--token={{API_TOKEN}}"]}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(
-        installations,
-        HashMap::from([("API_TOKEN".to_string(), raw_secret("real"))]),
-    );
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::CLIServer { args, .. } => {
-            assert_eq!(args, &vec!["--token=real".to_string()]);
-        }
-        other => panic!("expected CLI server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_command_config_preserves_literal_env_when_synthesizing_arg_placeholder() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","args":["--token={{API_TOKEN}}"],"env":{"LOG_LEVEL":"info"}}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(
-        installations,
-        HashMap::from([("API_TOKEN".to_string(), raw_secret("real"))]),
-    );
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::CLIServer { args, env, .. } => {
-            assert_eq!(args, &vec!["--token=real".to_string()]);
-            assert_eq!(env.get("LOG_LEVEL").map(String::as_str), Some("info"));
-        }
-        other => panic!("expected CLI server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_url_config_preserves_proxy_url_and_header() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"url":"https://proxy.example/mcp","headers":{"Authorization":"Bearer proxy-token"}}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(installations, HashMap::new());
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::SSEServer { url, headers } => {
-            assert_eq!(url, "https://proxy.example/mcp");
-            assert_eq!(
-                headers.get("Authorization").map(String::as_str),
-                Some("Bearer proxy-token")
-            );
-        }
-        other => panic!("expected SSE server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_url_config_preserves_header_despite_colliding_local_secret() {
-    // A server-rendered proxy header must not be overwritten by a local secret that
-    // happens to share the header's key name (`apply_secrets` implicit key-name match).
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"url":"https://proxy.example/mcp","headers":{"Authorization":"Bearer proxy-token"}}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(
-        installations,
-        HashMap::from([("Authorization".to_string(), raw_secret("local-secret"))]),
-    );
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::SSEServer { url, headers } => {
-            assert_eq!(url, "https://proxy.example/mcp");
-            assert_eq!(
-                headers.get("Authorization").map(String::as_str),
-                Some("Bearer proxy-token")
-            );
-        }
-        other => panic!("expected SSE server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_command_config_preserves_literal_env_despite_colliding_local_secret() {
-    // A literal env value rendered by the server must survive even when a local secret
-    // shares the env key name.
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","env":{"LOG_LEVEL":"info"}}}}"#,
-        None,
-        "github",
-    )
-    .unwrap();
-    let rendered = render_installations(
-        installations,
-        HashMap::from([("LOG_LEVEL".to_string(), raw_secret("debug"))]),
-    );
-
-    match &rendered["GitHub MCP"].transport_type {
-        JSONTransportType::CLIServer { env, .. } => {
-            assert_eq!(env.get("LOG_LEVEL").map(String::as_str), Some("info"));
-        }
-        other => panic!("expected CLI server, got {other:?}"),
-    }
-}
-
-#[test]
-fn managed_command_config_missing_secret_is_rejected_before_serialization() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","args":["--token={{API_TOKEN}}"]}}}"#,
-        None,
-        "github",
-    )
-    .expect("managed MCP config should parse");
-    let error = AgentDriver::mcp_installations_to_json(installations, &HashMap::new())
-        .expect_err("an unresolved secret must not reach the harness config");
-
-    match error {
-        AgentDriverError::MCPUnresolvedSecrets {
-            server_name,
-            secret_names,
-        } => {
-            assert_eq!(server_name, "GitHub MCP");
-            assert_eq!(secret_names, vec!["API_TOKEN".to_string()]);
-        }
-        other => panic!("expected unresolved MCP secrets, got {other:?}"),
-    }
-}
-
-#[test]
-fn ephemeral_mcp_missing_secret_is_filtered_before_spawn() {
-    let installations = AgentDriver::installations_from_managed_client_config_json(
-        r#"{"mcpServers":{"GitHub MCP":{"url":"https://example.com/mcp","headers":{"Authorization":"Bearer {{API_TOKEN}}"}}}}"#,
-        None,
-        "github",
-    )
-    .expect("managed MCP config should parse");
-    let (ready, failures) =
-        AgentDriver::apply_secrets_to_ephemeral_mcp_installations(installations, &HashMap::new());
-
-    assert!(ready.is_empty());
-    assert_eq!(
-        failures,
-        vec!["'GitHub MCP' was not started: unresolved secret reference(s): API_TOKEN".to_string()]
-    );
-}
-
-// ── Ephemeral MCP installation ids: stable across rebuilds ─────────────────
-
-#[test]
-fn ephemeral_installation_id_is_stable_across_resolutions_for_same_run() {
-    let task_id: AmbientAgentTaskId = "550e8400-e29b-41d4-a716-446655440010".parse().unwrap();
-    let config_json =
-        r#"{"mcpServers":{"slack":{"url":"https://app.warp.dev/mcp/integration-proxy/slack"}}}"#;
-
-    // Same run re-resolving the same server after a rebuild.
-    let first = AgentDriver::installations_from_managed_client_config_json(
-        config_json,
-        Some(task_id),
-        "slack",
-    )
-    .unwrap();
-    let second = AgentDriver::installations_from_managed_client_config_json(
-        config_json,
-        Some(task_id),
-        "slack",
-    )
-    .unwrap();
-
-    assert_eq!(first.len(), 1);
-    assert_eq!(second.len(), 1);
-    assert_eq!(
-        first[0].uuid(),
-        second[0].uuid(),
-        "same run + same server must yield the same id across rebuilds"
-    );
-}
-
-#[test]
-fn ephemeral_installation_id_differs_across_runs() {
-    let task_id_a: AmbientAgentTaskId = "550e8400-e29b-41d4-a716-446655440011".parse().unwrap();
-    let task_id_b: AmbientAgentTaskId = "550e8400-e29b-41d4-a716-446655440012".parse().unwrap();
-    let config_json =
-        r#"{"mcpServers":{"slack":{"url":"https://app.warp.dev/mcp/integration-proxy/slack"}}}"#;
-
-    let a = AgentDriver::installations_from_managed_client_config_json(
-        config_json,
-        Some(task_id_a),
-        "slack",
-    )
-    .unwrap();
-    let b = AgentDriver::installations_from_managed_client_config_json(
-        config_json,
-        Some(task_id_b),
-        "slack",
-    )
-    .unwrap();
-
-    assert_ne!(
-        a[0].uuid(),
-        b[0].uuid(),
-        "different runs must not collide onto the same id"
-    );
-}
-
-#[test]
-fn ephemeral_installation_id_differs_across_servers_in_same_run() {
-    let task_id: AmbientAgentTaskId = "550e8400-e29b-41d4-a716-446655440013".parse().unwrap();
-    let slack_config =
-        r#"{"mcpServers":{"slack":{"url":"https://app.warp.dev/mcp/integration-proxy/slack"}}}"#;
-    let linear_config =
-        r#"{"mcpServers":{"linear":{"url":"https://app.warp.dev/mcp/integration-proxy/linear"}}}"#;
-
-    let slack = AgentDriver::installations_from_managed_client_config_json(
-        slack_config,
-        Some(task_id),
-        "slack",
-    )
-    .unwrap();
-    let linear = AgentDriver::installations_from_managed_client_config_json(
-        linear_config,
-        Some(task_id),
-        "linear",
-    )
-    .unwrap();
-
-    assert_ne!(
-        slack[0].uuid(),
-        linear[0].uuid(),
-        "different servers in one run must not collide onto the same id"
-    );
-}
-
-#[test]
-fn ephemeral_installation_id_is_random_without_task_id() {
-    let config_json =
-        r#"{"mcpServers":{"slack":{"url":"https://app.warp.dev/mcp/integration-proxy/slack"}}}"#;
-
-    let first =
-        AgentDriver::installations_from_managed_client_config_json(config_json, None, "slack")
-            .unwrap();
-    let second =
-        AgentDriver::installations_from_managed_client_config_json(config_json, None, "slack")
-            .unwrap();
-
-    assert_ne!(
-        first[0].uuid(),
-        second[0].uuid(),
-        "no task_id means no rebuild to survive, so ids stay random"
-    );
-}
-
-// ── Built-in Factory MCP injection tests ────────────────────────────────────
-
-fn api_key_credentials() -> Credentials {
-    Credentials::ApiKey {
-        key: "wk-test-key".to_string(),
-        owner_type: None,
-    }
-}
-
-#[test]
-fn builtin_factory_mcp_attaches_with_api_key_credentials() {
-    let _flag = FeatureFlag::FactoryMcp.override_enabled(true);
-
-    let installation =
-        AgentDriver::builtin_factory_mcp_for_run(Some(&api_key_credentials()), &HashSet::new())
-            .expect("built-in Factory MCP should attach when eligible");
-
-    assert_eq!(installation.uuid(), FACTORY_MCP_INSTALLATION_UUID);
-    assert_eq!(
-        installation.templatable_mcp_server().name,
-        FACTORY_MCP_SERVER_NAME
-    );
-}
-
-#[test]
-fn builtin_factory_mcp_skipped_when_flag_disabled() {
-    let _flag = FeatureFlag::FactoryMcp.override_enabled(false);
-
-    assert!(
-        AgentDriver::builtin_factory_mcp_for_run(Some(&api_key_credentials()), &HashSet::new())
-            .is_none()
-    );
-}
-
-#[test]
-fn builtin_factory_mcp_skipped_without_credentials() {
-    let _flag = FeatureFlag::FactoryMcp.override_enabled(true);
-
-    assert!(AgentDriver::builtin_factory_mcp_for_run(None, &HashSet::new()).is_none());
-}
-
-#[test]
-fn builtin_factory_mcp_skipped_on_name_collision() {
-    let _flag = FeatureFlag::FactoryMcp.override_enabled(true);
-    // A user-configured server named `warp-factory` wins over the built-in.
-    let taken_server_names = HashSet::from([FACTORY_MCP_SERVER_NAME.to_string()]);
-
-    assert!(
-        AgentDriver::builtin_factory_mcp_for_run(Some(&api_key_credentials()), &taken_server_names)
-            .is_none()
-    );
-}
-
-#[test]
-fn managed_resolution_failure_includes_uid_and_message() {
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(1)
-        .returning(|_| Err(anyhow::anyhow!("not active")));
-
-    let err = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap_err();
-
-    match err {
-        AgentDriverError::ManagedMcpResolutionFailed { uid, message } => {
-            assert_eq!(uid, uuid);
-            assert!(message.contains("not active"));
-        }
-        other => panic!("expected managed MCP resolution failure, got {other:?}"),
-    }
-}
-
-fn transient_managed_mcp_error() -> anyhow::Error {
-    // A transport-level 503 with no GraphQL user-facing payload, matching what
-    // `send_graphql_request` produces for a genuinely transient backend failure (as opposed
-    // to a `UserFacingError`, which `ManagedMcpClient::create_managed_mcp_client_config`
-    // converts into a plain, untyped `anyhow!(message)`).
-    anyhow::Error::new(GraphQLError::HttpError {
-        status: StatusCode::SERVICE_UNAVAILABLE,
-        body: "unavailable".to_string(),
-    })
-}
-
-#[test]
-fn managed_resolution_retries_transient_error_then_succeeds() {
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let config_json =
-        r#"{"mcpServers":{"GitHub MCP":{"command":"npx","env":{"API_TOKEN":"{{API_TOKEN}}"}}}}"#;
-    let calls = Arc::new(AtomicUsize::new(0));
-    let calls_clone = Arc::clone(&calls);
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(2)
-        .returning(move |_| {
-            if calls_clone.fetch_add(1, Ordering::SeqCst) == 0 {
-                Err(transient_managed_mcp_error())
-            } else {
-                Ok(managed_client_config_output(config_json))
-            }
-        });
-
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert_eq!(resolved.ephemeral_installations.len(), 1);
-    assert_eq!(calls.load(Ordering::SeqCst), 2);
-}
-
-#[test]
-fn managed_resolution_does_not_retry_permanent_typed_http_error() {
-    // A typed but permanent (403) transport error must fail fast, same as the untyped
-    // user-facing error covered by `managed_resolution_failure_includes_uid_and_message`.
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(1)
-        .returning(|_| {
-            Err(anyhow::Error::new(GraphQLError::HttpError {
-                status: StatusCode::FORBIDDEN,
-                body: "forbidden".to_string(),
-            }))
-        });
-
-    let err = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap_err();
-
-    assert!(matches!(
-        err,
-        AgentDriverError::ManagedMcpResolutionFailed { uid, .. } if uid == uuid
-    ));
-}
-
-#[test]
-fn managed_resolution_exhausts_retry_budget_on_persistent_transient_error() {
-    let uuid = uuid::Uuid::parse_str("550e8400-e29b-41d4-a716-446655440000").unwrap();
-    let calls = Arc::new(AtomicUsize::new(0));
-    let calls_clone = Arc::clone(&calls);
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(MANAGED_MCP_RESOLVE_MAX_ATTEMPTS)
-        .returning(move |_| {
-            calls_clone.fetch_add(1, Ordering::SeqCst);
-            Err(transient_managed_mcp_error())
-        });
-
-    let err = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::Uuid(uuid)],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap_err();
-
-    assert!(matches!(
-        err,
-        AgentDriverError::ManagedMcpResolutionFailed { uid, .. } if uid == uuid
-    ));
-    // A persistently transient error must still exhaust the full retry budget rather than
-    // giving up early, and must not retry beyond it.
-    assert_eq!(
-        calls.load(Ordering::SeqCst),
-        MANAGED_MCP_RESOLVE_MAX_ATTEMPTS
-    );
-}
-
-#[test]
-fn well_known_resolution_retries_transient_error_then_succeeds() {
-    let _flag = FeatureFlag::WellKnownMcpIds.override_enabled(true);
-    let config_json = r#"{"mcpServers":{"linear":{"url":"https://app.warp.dev/mcp/integration-proxy/linear","headers":{"Authorization":"Bearer tok"}}}}"#;
-    let calls = Arc::new(AtomicUsize::new(0));
-    let calls_clone = Arc::clone(&calls);
-    let mut mock = MockManagedMcpClient::new();
-    mock.expect_create_managed_mcp_client_config()
-        .times(2)
-        .returning(move |_| {
-            if calls_clone.fetch_add(1, Ordering::SeqCst) == 0 {
-                Err(transient_managed_mcp_error())
-            } else {
-                Ok(managed_client_config_output(config_json))
-            }
-        });
-
-    // A transient failure must not silently skip the server the way a permanent one does
-    // (see `well_known_resolution_failure_skips_server`): it should retry and still resolve.
-    let resolved = block_on(AgentDriver::resolve_mcp_specs_with_local_uuids(
-        &[MCPSpec::WellKnown("linear".to_string())],
-        &HashSet::new(),
-        Arc::new(mock),
-        None,
-    ))
-    .unwrap();
-
-    assert_eq!(resolved.ephemeral_installations.len(), 1);
-    assert_eq!(calls.load(Ordering::SeqCst), 2);
-}
 
 // ── IdleTimeoutSender tests ──────────────────────────────────────────────────────
