+++ app/src/ai/mcp/file_mcp_watcher.rs
@@ -119,22 +119,49 @@ impl RepositorySubscriber for FileMCPSubscriber {
     }
 }
 
+/// A single source's currently-scheduled parse. Generation-tagged so a completion callback
+/// can tell whether it is still current or was superseded by a newer parse scheduled for the
+/// same source: `AbortHandle::abort` only takes effect the next time the background future is
+/// polled, so a superseded parse's completion can already be queued on the foreground executor
+/// by the time a replacement is scheduled. See
+/// [`FileMCPWatcher::update_servers_from_config_file`].
+struct InFlightParse {
+    generation: u64,
+    abort_handle: AbortHandle,
+}
+
+/// A batch of config sources whose first parse a consumer waits on before proceeding.
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+pub enum PendingScan {
+    /// Every provider config file in a cloud-environment repo, keyed by repo root.
+    CloudEnvRepo(PathBuf),
+    /// Every global (home-directory and Warp-managed) config source read at startup.
+    InitialGlobal,
+}
+
 /// Model that watches the filesystem for file-based MCP config changes and emits
 /// [`FileMCPWatcherEvent`]s.
 pub struct FileMCPWatcher {
     file_mcp_tx: Sender<FileMCPDetectionMessage>,
-    /// In-flight config parses keyed by source. Starting a replacement aborts
-    /// the previous parse so stale work cannot emit an event.
-    parse_abort_handles: HashMap<(PathBuf, MCPProvider), AbortHandle>,
+    /// The current in-flight parse for each source, if any. Starting a replacement removes
+    /// and aborts the previous record; only a completion callback whose captured generation
+    /// still matches the record here may act on its result. See
+    /// [`Self::update_servers_from_config_file`].
+    in_flight_parses: HashMap<(PathBuf, MCPProvider), InFlightParse>,
+    /// Monotonically increasing counter; each scheduled parse claims the next value as its
+    /// generation. Never reused, even after a source's record is removed, so a long-superseded
+    /// parse's callback can never appear current again by coincidence.
+    next_parse_generation: u64,
     /// Watcher handles for home provider subdirectories (e.g. `~/.codex`), keyed by subdir path.
     /// Used to cleanup watchers when the subdir is deleted at runtime.
     home_provider_watchers: HashMap<PathBuf, (ModelHandle<Repository>, SubscriberId)>,
     /// Set of project repository root paths we are already watching for file-based MCP configs.
     /// Used purely for deduplication — we never tear down project watchers during the session.
     project_repo_watchers: HashSet<PathBuf>,
-    /// Tracks how many provider config files remain to be parsed for each cloud environment repo.
-    /// When the count reaches zero, a `CloudEnvironmentScanComplete` event is emitted.
-    cloud_env_pending: HashMap<PathBuf, usize>,
+    /// Config sources each in-progress scan still owes a first result for. A source is settled
+    /// by its current parse completing or by an explicit removal; once a scan's set drains, its
+    /// entry is removed and [`FileMCPWatcherEvent::ScanComplete`] is emitted.
+    pending_scans: HashMap<PendingScan, HashSet<(PathBuf, MCPProvider)>>,
 }
 
 impl FileMCPWatcher {
@@ -157,9 +184,12 @@ impl FileMCPWatcher {
                 if should_watch_repository(*source, settings_mode) {
                     let repo_path = repository.as_ref(ctx).root_dir().to_local_path_lossy();
                     if matches!(source, RepoDetectionSource::CloudEnvironmentPrep) {
-                        let count =
-                            providers_in_scope(repo_path.clone(), repo_path.clone()).count();
-                        me.cloud_env_pending.insert(repo_path.clone(), count);
+                        me.pending_scans.insert(
+                            PendingScan::CloudEnvRepo(repo_path.clone()),
+                            providers_in_scope(repo_path.clone(), repo_path.clone())
+                                .map(|(provider, config_path)| (config_path, provider))
+                                .collect(),
+                        );
                     }
                     me.register_repo_for_file_mcp_watching(repo_path, ctx, file_mcp_tx.clone());
                 }
@@ -176,58 +206,59 @@ impl FileMCPWatcher {
         );
 
         let mut home_provider_watchers = HashMap::new();
-        let mut initial_config_parses = Vec::new();
-        if let Some(mcp_config_path) = warp_managed_mcp_config_path() {
-            initial_config_parses.push((
-                mcp_config_path.config_path,
-                mcp_config_path.root_path,
-                MCPProvider::Warp,
-            ));
-        }
-
         if let Some(home_dir) = dirs::home_dir() {
-            for provider in MCPProvider::iter() {
-                if provider == MCPProvider::Warp {
+            for provider in MCPProvider::iter().filter(|provider| *provider != MCPProvider::Warp) {
+                let Some(subdir) = home_subdir_to_watch(provider) else {
                     continue;
-                }
-                match home_subdir_to_watch(provider) {
-                    None => {
-                        // Initial scan of config files for providers whose config lives directly in
-                        // home (i.e. ~/.claude.json). HomeDirectoryWatcher handles incremental updates.
-                        let Some(config_path) = home_config_file_path(provider) else {
-                            continue;
-                        };
-                        initial_config_parses.push((config_path, home_dir.clone(), provider));
-                    }
-                    Some(subdir) => {
-                        // For providers whose home config lives in a subdir (e.g. ~/.codex for Codex)
-                        // start watching the subdir for file-based MCP servers, if it exists.
-                        let subdir_path = home_dir.join(&subdir);
-                        // Note: this will fail if the subdir doesn't exist yet.
-                        // We register upon creation of the subdir via HomeDirectoryWatcher.
-                        Self::watch_home_provider_dir(
-                            &subdir_path,
-                            home_dir.clone(),
-                            file_mcp_tx.clone(),
-                            &mut home_provider_watchers,
-                            ctx,
-                        );
-                    }
-                }
+                };
+                // Note: this will fail if the subdir doesn't exist yet.
+                // We register upon creation of the subdir via HomeDirectoryWatcher.
+                Self::watch_home_provider_dir(
+                    &home_dir.join(subdir),
+                    home_dir.clone(),
+                    file_mcp_tx.clone(),
+                    &mut home_provider_watchers,
+                    ctx,
+                );
             }
         }
 
+        let global_configs = global_config_sources();
         let mut watcher = Self {
             file_mcp_tx,
-            parse_abort_handles: HashMap::new(),
+            in_flight_parses: HashMap::new(),
+            next_parse_generation: 0,
             home_provider_watchers,
             project_repo_watchers: HashSet::new(),
-            cloud_env_pending: HashMap::new(),
+            pending_scans: HashMap::from([(
+                PendingScan::InitialGlobal,
+                global_configs
+                    .iter()
+                    .map(|(config_path, _, provider)| (config_path.clone(), *provider))
+                    .collect(),
+            )]),
         };
-        for (config_path, root_path, provider) in initial_config_parses {
+        // Every global source is read explicitly so the initial-global scan settles even when a
+        // provider's home subdir watcher never registers. A subdir watcher's own scan of the
+        // same file simply supersedes the explicit parse.
+        for (config_path, root_path, provider) in global_configs {
             watcher.update_servers_from_config_file(&config_path, root_path, provider, ctx);
         }
         watcher
+    }
+
+    /// A watcher with no filesystem subscriptions and no owed scans, so tests can feed
+    /// consumers a deterministic event sequence.
+    #[cfg(test)]
+    pub(crate) fn new_inert() -> Self {
+        Self {
+            file_mcp_tx: async_channel::unbounded().0,
+            in_flight_parses: HashMap::new(),
+            next_parse_generation: 0,
+            home_provider_watchers: HashMap::new(),
+            project_repo_watchers: HashSet::new(),
+            pending_scans: HashMap::new(),
+        }
     }
 
     #[cfg(feature = "tui")]
@@ -297,7 +328,6 @@ impl FileMCPWatcher {
         home_provider_watchers: &mut HashMap<PathBuf, (ModelHandle<Repository>, SubscriberId)>,
         ctx: &mut ModelContext<Self>,
     ) {
-        // If the subdir is already being watched, return early.
         if home_provider_watchers.contains_key(subdir_path) {
             return;
         }
@@ -320,7 +350,6 @@ impl FileMCPWatcher {
                 return;
             }
         };
-
         let subscriber = Box::new(FileMCPSubscriber {
             stored_dir: home_dir,
             message_tx: file_mcp_tx,
@@ -420,12 +449,7 @@ impl FileMCPWatcher {
                             repo_handle.update(ctx, |repo, ctx| repo.stop_watching(id, ctx));
                         }
                         let config_path = home_dir.join(provider.home_config_path());
-                        self.abort_config_parse(&config_path, provider);
-                        ctx.emit(FileMCPWatcherEvent::ConfigRemoved {
-                            config_path,
-                            root_path: home_dir.clone(),
-                            provider,
-                        });
+                        self.remove_config(config_path, home_dir.clone(), provider, ctx);
                     }
                 }
             }
@@ -543,29 +567,88 @@ impl FileMCPWatcher {
         // update. Parse the replacement without transiently removing the
         // last-known-good servers.
         if was_deleted && !was_added {
-            self.abort_config_parse(&config_path, provider);
-            ctx.emit(FileMCPWatcherEvent::ConfigRemoved {
-                config_path: config_path.clone(),
-                root_path: root_path.clone(),
-                provider,
-            });
-        }
-        if was_added {
+            self.remove_config(config_path, root_path, provider, ctx);
+        } else if was_added {
             self.update_servers_from_config_file(&config_path, root_path, provider, ctx);
         }
     }
 
+    /// Drops any in-flight parse for the source, reports it removed, and settles it for any
+    /// pending scan still waiting on it.
+    fn remove_config(
+        &mut self,
+        config_path: PathBuf,
+        root_path: PathBuf,
+        provider: MCPProvider,
+        ctx: &mut ModelContext<Self>,
+    ) {
+        self.abort_config_parse(&config_path, provider);
+        ctx.emit(FileMCPWatcherEvent::ConfigRemoved {
+            config_path: config_path.clone(),
+            root_path,
+            provider,
+        });
+        self.settle_pending_source(&config_path, provider, ctx);
+    }
+
+    /// Marks the source as no longer owed by any pending scan, emitting
+    /// [`FileMCPWatcherEvent::ScanComplete`] for each scan this drains.
+    fn settle_pending_source(
+        &mut self,
+        config_path: &Path,
+        provider: MCPProvider,
+        ctx: &mut ModelContext<Self>,
+    ) {
+        let source = (config_path.to_path_buf(), provider);
+        let mut completed = Vec::new();
+        self.pending_scans.retain(|scan, sources| {
+            sources.remove(&source);
+            if sources.is_empty() {
+                completed.push(scan.clone());
+            }
+            !sources.is_empty()
+        });
+        for scan in completed {
+            ctx.emit(FileMCPWatcherEvent::ScanComplete(scan));
+        }
+    }
+
     fn abort_config_parse(&mut self, config_path: &Path, provider: MCPProvider) {
-        if let Some(abort_handle) = self
-            .parse_abort_handles
+        if let Some(in_flight) = self
+            .in_flight_parses
             .remove(&(config_path.to_path_buf(), provider))
         {
-            abort_handle.abort();
+            in_flight.abort_handle.abort();
+        }
+    }
+
+    /// Returns whether the parse identified by `key`/`generation` is still the source's
+    /// current one, removing its record if so. A `false` result means it was superseded by a
+    /// newer parse scheduled for the same source (or already explicitly removed) since it was
+    /// scheduled: `AbortHandle::abort` only takes effect the next time the background future is
+    /// polled, so a superseded parse's completion callback can still run after that. The
+    /// caller must treat a `false` result as fully stale because its possibly outdated result
+    /// would preempt whatever superseded it.
+    fn take_current_in_flight_parse(
+        &mut self,
+        key: &(PathBuf, MCPProvider),
+        generation: u64,
+    ) -> bool {
+        match self.in_flight_parses.get(key) {
+            Some(in_flight) if in_flight.generation == generation => {
+                self.in_flight_parses.remove(key);
+                true
+            }
+            _ => false,
         }
     }
 
     /// Asynchronously reads and parses the MCP configuration file at `config_file_path`,
     /// then emits a [`FileMCPWatcherEvent::ConfigParsed`] event.
+    ///
+    /// Every call for a given `(config_file_path, provider)` source claims a fresh generation
+    /// (see [`Self::take_current_in_flight_parse`]) and aborts any in-flight parse for that
+    /// source.
     fn update_servers_from_config_file(
         &mut self,
         config_file_path: &Path,
@@ -577,25 +660,31 @@ impl FileMCPWatcher {
         let key = (config_file_path.clone(), provider);
         let callback_key = key.clone();
         self.abort_config_parse(&config_file_path, provider);
+
+        let generation = self.next_parse_generation;
+        self.next_parse_generation += 1;
+
         let parse = ctx.spawn(
             async move { parse_mcp_config_file(&config_file_path, provider).await },
             move |me, outcome, ctx| {
-                me.parse_abort_handles.remove(&callback_key);
-                let repo_path_for_countdown = root_path.clone();
-                emit_parse_outcome(outcome, callback_key.0.clone(), root_path, provider, ctx);
-                if let Some(count) = me.cloud_env_pending.get_mut(&repo_path_for_countdown) {
-                    *count = count.saturating_sub(1);
-                    if *count == 0 {
-                        // If we've parsed all MCP config files for the cloud environment repo, emit a `CloudEnvironmentScanComplete` event.
-                        me.cloud_env_pending.remove(&repo_path_for_countdown);
-                        ctx.emit(FileMCPWatcherEvent::CloudEnvMcpScanComplete {
-                            repo_path: repo_path_for_countdown,
-                        });
-                    }
+                if !me.take_current_in_flight_parse(&callback_key, generation) {
+                    // Superseded: a newer parse for this source already owns the record (or
+                    // an explicit removal already settled it). Do not touch anything else.
+                    return;
                 }
+
+                let (config_path, provider) = callback_key;
+                emit_parse_outcome(outcome, config_path.clone(), root_path, provider, ctx);
+                me.settle_pending_source(&config_path, provider, ctx);
             },
         );
-        self.parse_abort_handles.insert(key, parse.abort_handle());
+        self.in_flight_parses.insert(
+            key,
+            InFlightParse {
+                generation,
+                abort_handle: parse.abort_handle(),
+            },
+        );
     }
 }
 
@@ -629,6 +718,27 @@ fn config_change_flags(update: &RepositoryUpdate, config_path: &Path) -> (bool, 
         || update.moved.keys().any(|file| file.path == config_path);
     (was_deleted, was_added)
 }
+
+/// Every global config source as `(config_path, root_path, provider)`: the Warp-managed config
+/// plus each other provider's home-directory config.
+fn global_config_sources() -> Vec<(PathBuf, PathBuf, MCPProvider)> {
+    let mut sources = Vec::new();
+    if let Some(config) = warp_managed_mcp_config_path() {
+        sources.push((config.config_path, config.root_path, MCPProvider::Warp));
+    }
+    if let Some(home_dir) = dirs::home_dir() {
+        sources.extend(
+            MCPProvider::iter()
+                .filter(|provider| *provider != MCPProvider::Warp)
+                .filter_map(|provider| {
+                    home_config_file_path(provider)
+                        .map(|config_path| (config_path, home_dir.clone(), provider))
+                }),
+        );
+    }
+    sources
+}
+
 /// Returns an iterator of `(provider, config_path)` pairs for MCP providers whose configuration file
 /// paths fall within the watched directory.
 fn providers_in_scope(
@@ -843,8 +953,9 @@ pub enum FileMCPWatcherEvent {
     },
     /// A config could not be read or parsed. Consumers should preserve the last-known-good state.
     ConfigError { diagnostic: FileMCPConfigDiagnostic },
-    /// All provider config files for a cloud environment repo have been parsed.
-    CloudEnvMcpScanComplete { repo_path: PathBuf },
+    /// Every source owed by the scan has been parsed or removed. Emitted after the events
+    /// carrying those sources' results.
+    ScanComplete(PendingScan),
 }
 
 impl Entity for FileMCPWatcher {
